test: larger and non-empty preference scenarios
