+++ erpc/networks_consensus_test.go
@@ -3,7 +3,6 @@ package erpc
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"strings"
@@ -22,4737 +21,2162 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-var (
-	successResponse, _ = common.NewJsonRpcResponse(1, "0x7a", nil)
-)
+type consensusTestCase struct {
+	name                 string
+	description          string
+	upstreams            []*common.UpstreamConfig
+	consensusConfig      *common.ConsensusPolicyConfig
+	retryPolicy          *common.RetryPolicyConfig
+	timeoutPolicy        *common.TimeoutPolicyConfig
+	circuitBreakerPolicy *common.CircuitBreakerPolicyConfig
+	hedgePolicy          *common.HedgePolicyConfig
+	mockResponses        []mockResponse
+	expectedCalls        []int
+	expectedResult       *expectedResult
+	expectedError        *expectedError
+	expectedPendingMocks int
+	setupFn              func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry)
+	requestMethod        string
+	requestParams        []interface{}
+}
+
+type mockResponse struct {
+	status int
+	body   map[string]interface{}
+	delay  time.Duration
+}
+
+type expectedResult struct {
+	contains      string
+	check         func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration)
+	jsonRpcResult string
+}
+
+type expectedError struct {
+	code     common.ErrorCode
+	contains string
+}
 
 func init() {
 	util.ConfigureTestLogger()
 }
 
-func TestNetwork_ConsensusPolicy(t *testing.T) {
-	tests := []struct {
-		name                    string
-		upstreams               []*common.UpstreamConfig
-		hasRetries              bool
-		request                 map[string]interface{}
-		mockResponses           []map[string]interface{}
-		maxParticipants         int
-		agreementThreshold      *int
-		disputeBehavior         *common.ConsensusDisputeBehavior
-		lowParticipantsBehavior *common.ConsensusLowParticipantsBehavior
-		expectedCalls           []int // Number of expected calls for each upstream
-		expectedResponse        *common.NormalizedResponse
-		expectedError           *common.ErrorCode
-		expectedMsg             *string
-	}{
-		// {
-		// 	name:            "agreed_unknown_error_response_on_upstreams",
-		// 	maxParticipants: 3,
-		// 	upstreams: []*common.UpstreamConfig{
-		// 		{
-		// 			Id:       "test1",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc1-failure.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test2",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc2-failure.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test3",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc3-failure.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 	},
-		// 	request: map[string]interface{}{
-		// 		"method": "eth_chainId",
-		// 		"params": []interface{}{},
-		// 	},
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"error": map[string]interface{}{
-		// 				"code":    -32000,
-		// 				"message": "random error",
-		// 			},
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"error": map[string]interface{}{
-		// 				"code":    -32000,
-		// 				"message": "random error",
-		// 			},
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"error": map[string]interface{}{
-		// 				"code":    -32000,
-		// 				"message": "random error",
-		// 			},
-		// 		},
-		// 	},
-		// 	expectedCalls: []int{1, 1, 1}, // Each upstream called once
-		// 	expectedError: pointer(common.ErrCodeUpstreamRequest),
-		// 	expectedMsg:   pointer("random error"),
-		// },
+func TestConsensusPolicy(t *testing.T) {
+	tests := []consensusTestCase{
 		{
-			name:               "some_participants_return_error_succes_is_below_threshold",
-			maxParticipants:    3,
-			agreementThreshold: pointer(3),
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
+			name:        "only_block_head_leader_dispute_selects_leader_non_error",
+			description: "Dispute with OnlyBlockHeadLeader: pick leader's non-error result",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:    3,
+				AgreementThreshold: 3, // ensure dispute path (no group meets threshold, not low participants)
+				DisputeBehavior:    common.ConsensusDisputeBehaviorOnlyBlockHeadLeader,
 			},
-			request: map[string]interface{}{
-				"method": "eth_chainId",
-				"params": []interface{}{},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xleader")}, // leader upstream
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+				{status: 200, body: jsonRpcSuccess("0xbbb")},
 			},
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "internal error",
-					},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0xleader"`},
+			setupFn: func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry) {
+				ups := reg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
+				// Set unique leader: upstream 1 has highest latest block
+				ups[0].EvmStatePoller().SuggestLatestBlock(300)
+				ups[1].EvmStatePoller().SuggestLatestBlock(100)
+				ups[2].EvmStatePoller().SuggestLatestBlock(100)
 			},
-			expectedCalls: []int{1, 1, 1},
-			expectedError: pointer(common.ErrCodeConsensusLowParticipants),
-			expectedMsg:   pointer("not enough participants"),
 		},
 		{
-			name:               "some_participants_return_error_success_is_above_threshold",
-			maxParticipants:    3,
-			agreementThreshold: pointer(2),
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
+			name:        "prefer_block_head_leader_dispute_prefers_leader_non_error",
+			description: "Dispute with PreferBlockHeadLeader: prefer leader's non-error result",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:    3,
+				AgreementThreshold: 3,
+				DisputeBehavior:    common.ConsensusDisputeBehaviorPreferBlockHeadLeader,
 			},
-			request: map[string]interface{}{
-				"method": "eth_chainId",
-				"params": []interface{}{},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xleader")},
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+				{status: 200, body: jsonRpcSuccess("0xbbb")},
 			},
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "internal error",
-					},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0xleader"`},
+			setupFn: func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry) {
+				ups := reg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
+				ups[0].EvmStatePoller().SuggestLatestBlock(300)
+				ups[1].EvmStatePoller().SuggestLatestBlock(100)
+				ups[2].EvmStatePoller().SuggestLatestBlock(100)
 			},
-			expectedCalls: []int{1, 1, 1},
-			expectedResponse: common.NewNormalizedResponse().
-				WithJsonRpcResponse(successResponse),
 		},
 		{
-			name:                    "some_participants_return_error_and_return_error_on_low_participants",
-			maxParticipants:         3,
-			agreementThreshold:      pointer(3),
-			lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorReturnError),
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
+			name:        "prefer_block_head_leader_leader_error_fallback_to_consensus",
+			description: "PreferBlockHeadLeader with leader error falls back to accept-most-common logic",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:    3,
+				AgreementThreshold: 3, // dispute path
+				DisputeBehavior:    common.ConsensusDisputeBehaviorPreferBlockHeadLeader,
 			},
-			request: map[string]interface{}{
-				"method": "eth_chainId",
-				"params": []interface{}{},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "leader error")}, // leader
+				{status: 200, body: jsonRpcSuccess("0xagreed")},           // best non-empty group (count 2)
+				{status: 200, body: jsonRpcSuccess("0xagreed")},
 			},
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "internal error",
-					},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0xagreed"`},
+			setupFn: func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry) {
+				ups := reg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
+				ups[0].EvmStatePoller().SuggestLatestBlock(300)
+				ups[1].EvmStatePoller().SuggestLatestBlock(100)
+				ups[2].EvmStatePoller().SuggestLatestBlock(100)
 			},
-			expectedCalls: []int{1, 1, 1},
-			expectedError: pointer(common.ErrCodeConsensusLowParticipants),
-			expectedMsg:   pointer("not enough participants"),
 		},
 		{
-			name:                    "some_participants_return_error_but_accept_most_common_valid_result",
-			maxParticipants:         3,
-			agreementThreshold:      pointer(2),
-			lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult),
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
+			name:        "only_block_head_leader_low_participants_selects_leader",
+			description: "Low participants with OnlyBlockHeadLeader selects leader's non-error result",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      3, // force low participants
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorOnlyBlockHeadLeader,
 			},
-			request: map[string]interface{}{
-				"method": "eth_chainId",
-				"params": []interface{}{},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xleader")},
+				{status: 200, body: jsonRpcError(-32603, "infra")},
 			},
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "internal error",
-					},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
+			expectedCalls:  []int{1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0xleader"`},
+			setupFn: func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry) {
+				ups := reg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
+				ups[0].EvmStatePoller().SuggestLatestBlock(300)
+				ups[1].EvmStatePoller().SuggestLatestBlock(100)
 			},
-			expectedCalls: []int{1, 1, 1},
-			expectedResponse: common.NewNormalizedResponse().
-				WithJsonRpcResponse(successResponse),
 		},
 		{
-			name:                    "not_enough_upstreams_low_participants_behavior_accept_most_common_valid_result",
-			maxParticipants:         3,
-			agreementThreshold:      pointer(3),
-			lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult),
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-failure.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
+			name:        "only_block_head_leader_low_participants_no_leader_non_error_errors",
+			description: "Low participants with OnlyBlockHeadLeader and no leader non-error -> low participants error",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      3,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorOnlyBlockHeadLeader,
 			},
-			request: map[string]interface{}{
-				"method": "eth_chainId",
-				"params": []interface{}{},
-			},
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x7a",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "internal error",
-					},
-				},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "leader error")}, // leader error only
+				{status: 200, body: jsonRpcSuccess("0xother")},
 			},
 			expectedCalls: []int{1, 1},
-			expectedResponse: common.NewNormalizedResponse().
-				WithJsonRpcResponse(successResponse),
-		},
-		// {
-		// 	name:                    "only_block_head_leader_selects_highest_block_upstream",
-		// 	agreementThreshold:      pointer(2),
-		// 	maxParticipants:         3,
-		// 	lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorOnlyBlockHeadLeader),
-		// 	disputeBehavior:         pointer(common.ConsensusDisputeBehaviorOnlyBlockHeadLeader),
-		// 	upstreams: []*common.UpstreamConfig{
-		// 		{
-		// 			Id:       "test1",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc1-leader.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test2",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc2-leader.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test3",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc3-leader.localhost", // This will be the leader
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 	},
-		// 	request: map[string]interface{}{
-		// 		"method": "eth_chainId",
-		// 		"params": []interface{}{},
-		// 	},
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x5a",
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x6a",
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a",
-		// 		},
-		// 	},
-		// 	expectedCalls: []int{
-		// 		1, // test1 should NOT be used
-		// 		1, // test2 should NOT be used
-		// 		1, // test3 (leader) should be used
-		// 	},
-		// 	expectedResponse: common.NewNormalizedResponse().
-		// 		WithJsonRpcResponse(successResponse),
-		// },
-		// {
-		// 	name:                    "prefer_block_head_leader_includes_leader_in_participants",
-		// 	agreementThreshold:      pointer(2),
-		// 	maxParticipants:         3,
-		// 	lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorPreferBlockHeadLeader),
-		// 	upstreams: []*common.UpstreamConfig{
-		// 		{
-		// 			Id:       "test1",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc1-leader.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test2",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc2-leader.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test3",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc3-leader.localhost", // This will be the leader
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 	},
-		// 	request: map[string]interface{}{
-		// 		"method": "eth_chainId",
-		// 		"params": []interface{}{},
-		// 	},
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a",
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a", // Same result achieves consensus
-		// 		},
-		// 		// test3 not mocked since it won't be called due to early consensus
-		// 	},
-		// 	expectedCalls: []int{
-		// 		1, // test1 should be used
-		// 		1, // test2 should be used (both return 0x7a, achieving consensus)
-		// 		0, // test3 (leader) not called due to early consensus
-		// 	},
-		// 	expectedResponse: common.NewNormalizedResponse().
-		// 		WithJsonRpcResponse(successResponse),
-		// },
-		// {
-		// 	name:                    "only_block_head_leader_no_leader_available_uses_normal_selection",
-		// 	maxParticipants:         2,
-		// 	lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorOnlyBlockHeadLeader),
-		// 	upstreams: []*common.UpstreamConfig{
-		// 		{
-		// 			Id:       "test1",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc1-no-leader.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test2",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc2-no-leader.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 	},
-		// 	request: map[string]interface{}{
-		// 		"method": "eth_chainId",
-		// 		"params": []interface{}{},
-		// 	},
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a",
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a",
-		// 		},
-		// 	},
-		// 	expectedCalls: []int{
-		// 		1, // test1 should be called (normal selection)
-		// 		1, // test2 should be called (normal selection)
-		// 	},
-		// 	expectedResponse: common.NewNormalizedResponse().
-		// 		WithJsonRpcResponse(successResponse),
-		// },
-		// {
-		// 	name:                    "low_participants_with_prefer_block_head_leader_fallback",
-		// 	maxParticipants:         3,
-		// 	agreementThreshold:      pointer(2),
-		// 	lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorPreferBlockHeadLeader),
-		// 	upstreams: []*common.UpstreamConfig{
-		// 		{
-		// 			Id:       "test1",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc1-fallback.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test2",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc2-fallback.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 	},
-		// 	request: map[string]interface{}{
-		// 		"method": "eth_chainId",
-		// 		"params": []interface{}{},
-		// 	},
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a",
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a",
-		// 		},
-		// 	},
-		// 	expectedCalls: []int{
-		// 		1, // test1 should be called
-		// 		1, // test2 should be called
-		// 	},
-		// 	expectedResponse: common.NewNormalizedResponse().
-		// 		WithJsonRpcResponse(successResponse),
-		// },
-		// {
-		// 	name:                    "low_participants_with_prefer_block_head_leader",
-		// 	maxParticipants:         3,
-		// 	agreementThreshold:      pointer(4),
-		// 	lowParticipantsBehavior: pointer(common.ConsensusLowParticipantsBehaviorPreferBlockHeadLeader),
-		// 	upstreams: []*common.UpstreamConfig{
-		// 		{
-		// 			Id:       "test1",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc1-leader.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test2",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc2-follower.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 		{
-		// 			Id:       "test3",
-		// 			Type:     common.UpstreamTypeEvm,
-		// 			Endpoint: "http://rpc3-follower.localhost",
-		// 			Evm: &common.EvmUpstreamConfig{
-		// 				ChainId: 123,
-		// 			},
-		// 		},
-		// 	},
-		// 	request: map[string]interface{}{
-		// 		"method": "eth_chainId",
-		// 		"params": []interface{}{},
-		// 	},
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7a",
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0x7b", // test2 response (needed to detect low participants)
-		// 		},
-		// 		// test3 not mocked since it won't be called due to block head leader short-circuit
-		// 	},
-		// 	expectedCalls: []int{
-		// 		1, // test1 (leader) should be called and its result used
-		// 		1, // test2 (needed to detect low participants)
-		// 		0, // test3 (not called due to block head leader short-circuit)
-		// 	},
-		// 	expectedResponse: common.NewNormalizedResponse().
-		// 		WithJsonRpcResponse(successResponse),
-		// },
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-			defer util.AssertNoPendingMocks(t, 0)
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer func() {
-				cancel()
-				// Allow any hedge/consensus requests to complete before cleanup
-				time.Sleep(50 * time.Millisecond)
-			}()
-
-			// Setup network with consensus policy
-			mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
-
-			vr := thirdparty.NewVendorsRegistry()
-			pr, err := thirdparty.NewProvidersRegistry(
-				&log.Logger,
-				vr,
-				[]*common.ProviderConfig{},
-				nil,
-			)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
-				Connector: &common.ConnectorConfig{
-					Driver: "memory",
-					Memory: &common.MemoryConnectorConfig{
-						MaxItems:     100_000,
-						MaxTotalSize: "1MB",
-					},
-				},
-			})
-			if err != nil {
-				panic(err)
-			}
-
-			upsReg := upstream.NewUpstreamsRegistry(
-				ctx,
-				&log.Logger,
-				"prjA",
-				tt.upstreams,
-				ssr,
-				nil,
-				vr,
-				pr,
-				nil,
-				mt,
-				1*time.Second,
-			)
-
-			var retryPolicy *common.RetryPolicyConfig
-			if tt.hasRetries {
-				retryPolicy = &common.RetryPolicyConfig{
-					MaxAttempts: len(tt.upstreams),
-					Delay:       common.Duration(0),
-				}
-			}
-
-			agreementThreshold := 2
-			if tt.agreementThreshold != nil {
-				agreementThreshold = *tt.agreementThreshold
-			}
-
-			var lowParticipantsBehavior common.ConsensusLowParticipantsBehavior
-			if tt.lowParticipantsBehavior != nil {
-				lowParticipantsBehavior = *tt.lowParticipantsBehavior
-			}
-
-			disputeBehavior := common.ConsensusDisputeBehaviorReturnError
-			if tt.disputeBehavior != nil {
-				disputeBehavior = *tt.disputeBehavior
-			}
-
-			// Create consensus config and apply defaults
-			consensusConfig := &common.ConsensusPolicyConfig{
-				MaxParticipants:         tt.maxParticipants,
-				AgreementThreshold:      agreementThreshold,
-				DisputeBehavior:         disputeBehavior,
-				LowParticipantsBehavior: lowParticipantsBehavior,
-				PunishMisbehavior:       &common.PunishMisbehaviorConfig{},
-			}
-			// Apply defaults when behavior is not explicitly set
-			if tt.lowParticipantsBehavior == nil {
-				err := consensusConfig.SetDefaults()
-				if err != nil {
-					t.Fatal(err)
-				}
-			}
-
-			ntw, err := NewNetwork(
-				ctx,
-				&log.Logger,
-				"prjA",
-				&common.NetworkConfig{
-					Architecture: common.ArchitectureEvm,
-					Evm: &common.EvmNetworkConfig{
-						ChainId: 123,
-					},
-					Failsafe: []*common.FailsafeConfig{
-						{
-							MatchMethod: "*", // Match all methods to ensure consensus executor is selected
-							Retry:       retryPolicy,
-							Consensus:   consensusConfig,
-						},
-					},
-				},
-				nil,
-				upsReg,
-				mt,
-			)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			err = upsReg.Bootstrap(ctx)
-			if err != nil {
-				t.Fatal(err)
-			}
-			err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(123))
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			upstream.ReorderUpstreams(upsReg)
-
-			// Set up block numbers for leader selection tests
-			if tt.name == "only_block_head_leader_selects_highest_block_upstream" ||
-				tt.name == "prefer_block_head_leader_includes_leader_in_participants" {
-				// Get upstreams and set their block numbers
-				upsList := upsReg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
-				if len(upsList) >= 3 {
-					// Set block numbers: test3 will have the highest block
-					upsList[0].EvmStatePoller().SuggestLatestBlock(100) // test1
-					upsList[1].EvmStatePoller().SuggestLatestBlock(200) // test2
-					upsList[2].EvmStatePoller().SuggestLatestBlock(300) // test3 (leader)
-				}
-			} else if tt.name == "only_block_head_leader_no_leader_available_uses_normal_selection" {
-				// For the no-leader test, set all block numbers to 0
-				upsList := upsReg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
-				for _, ups := range upsList {
-					ups.EvmStatePoller().SuggestLatestBlock(0)
-				}
-			} else if tt.name == "low_participants_with_prefer_block_head_leader_fallback" {
-				// For the fallback test, set all upstreams to same block number (no clear leader)
-				upsList := upsReg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
-				for _, ups := range upsList {
-					ups.EvmStatePoller().SuggestLatestBlock(100) // All at same block
-				}
-			} else if tt.name == "low_participants_with_prefer_block_head_leader" {
-				// For the leader test, set test1 as leader with higher block
-				upsList := upsReg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
-				if len(upsList) >= 3 {
-					upsList[0].EvmStatePoller().SuggestLatestBlock(200) // test1 (leader)
-					upsList[1].EvmStatePoller().SuggestLatestBlock(100) // test2 (follower)
-					upsList[2].EvmStatePoller().SuggestLatestBlock(50)  // test3 (follower)
-				}
-			}
-
-			// Setup mock responses with expected call counts
-			for i, upstream := range tt.upstreams {
-				if tt.expectedCalls[i] > 0 {
-					gock.New(upstream.Endpoint).
-						Post("/").
-						Times(tt.expectedCalls[i]).
-						Reply(200).
-						SetHeader("Content-Type", "application/json").
-						JSON(tt.mockResponses[i])
-				}
-			}
-
-			// Make request
-			reqBytes, err := json.Marshal(tt.request)
-			if err != nil {
-				require.NoError(t, err)
-			}
-
-			fakeReq := common.NewNormalizedRequest(reqBytes)
-
-			// Log the initial request state
-			t.Logf("Initial request ID: %v, ptr: %p", fakeReq.ID(), fakeReq)
-
-			resp, err := ntw.Forward(ctx, fakeReq)
-
-			// Log the error for debugging
-			if err != nil {
-				t.Logf("Got error from Forward: %v", err)
-			} else if resp != nil && resp.Upstream() != nil {
-				t.Logf("Got response from upstream: %s", resp.Upstream().Id())
-			}
-
-			if tt.expectedError != nil {
-				assert.Error(t, err, "expected error but got nil")
-				assert.True(t, common.HasErrorCode(err, *tt.expectedError), "expected error code %s, got %s", *tt.expectedError, err)
-				if err != nil {
-					assert.Contains(t, err.Error(), *tt.expectedMsg, "expected error message %s, got %s", *tt.expectedMsg, err.Error())
-				}
-				assert.Nil(t, resp, "expected nil response")
-			} else {
-				assert.NoError(t, err)
-				assert.NotNil(t, resp)
-
-				expectedJrr, err := tt.expectedResponse.JsonRpcResponse()
-				assert.NoError(t, err)
-				assert.NotNil(t, expectedJrr)
-
-				actualJrr, err := resp.JsonRpcResponse()
-				assert.NoError(t, err)
-				assert.NotNil(t, actualJrr)
-
-				// Ensure both are either nil or non-nil
-				if expectedJrr != nil && actualJrr == nil {
-					t.Fatal("expected a JSON-RPC response but got nil")
-				} else if expectedJrr == nil && actualJrr != nil {
-					t.Fatal("expected no JSON-RPC response but got one")
-				} else if expectedJrr != nil && actualJrr != nil {
-					// Both are non-nil, compare the results
-					assert.Equal(t, string(expectedJrr.Result), string(actualJrr.Result))
-				}
-			}
-		})
-	}
-}
-
-func TestNetwork_Consensus_RetryIntermittentErrors(t *testing.T) {
-	util.ResetGock()
-	defer util.ResetGock()
-	util.SetupMocksForEvmStatePoller()
-	defer util.AssertNoPendingMocks(t, 0)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer func() {
-		cancel()
-		// Allow any hedge/consensus requests to complete before cleanup
-		time.Sleep(500 * time.Millisecond)
-	}()
-
-	// Setup network with consensus policy
-	mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
-
-	vr := thirdparty.NewVendorsRegistry()
-	pr, err := thirdparty.NewProvidersRegistry(
-		&log.Logger,
-		vr,
-		[]*common.ProviderConfig{},
-		nil,
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
-		Connector: &common.ConnectorConfig{
-			Driver: "memory",
-			Memory: &common.MemoryConnectorConfig{
-				MaxItems:     100_000,
-				MaxTotalSize: "1MB",
-			},
-		},
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	upsReg := upstream.NewUpstreamsRegistry(
-		ctx,
-		&log.Logger,
-		"prjA",
-		[]*common.UpstreamConfig{
-			{
-				Id:       "test1",
-				Type:     common.UpstreamTypeEvm,
-				Endpoint: "http://rpc1-dispute.localhost",
-				Evm: &common.EvmUpstreamConfig{
-					ChainId: 123,
-				},
-			},
-			{
-				Id:       "test2",
-				Type:     common.UpstreamTypeEvm,
-				Endpoint: "http://rpc2-dispute.localhost",
-				Evm: &common.EvmUpstreamConfig{
-					ChainId: 123,
-				},
-			},
-		},
-		ssr,
-		nil,
-		vr,
-		pr,
-		nil,
-		mt,
-		1*time.Second,
-	)
-
-	ntw, err := NewNetwork(
-		ctx,
-		&log.Logger,
-		"prjA",
-		&common.NetworkConfig{
-			Architecture: common.ArchitectureEvm,
-			Evm: &common.EvmNetworkConfig{
-				ChainId: 123,
-			},
-			Failsafe: []*common.FailsafeConfig{
-				{
-					MatchMethod: "*", // Match all methods to ensure consensus executor is selected
-					Retry: &common.RetryPolicyConfig{
-						MaxAttempts: 2,
-						Delay:       common.Duration(0),
-					},
-					Consensus: &common.ConsensusPolicyConfig{
-						MaxParticipants:         2,
-						AgreementThreshold:      2,
-						DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-						LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
-						PunishMisbehavior:       &common.PunishMisbehaviorConfig{},
-					},
-				},
-			},
-		},
-		nil,
-		upsReg,
-		mt,
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	err = upsReg.Bootstrap(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(123))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	upstream.ReorderUpstreams(upsReg)
-
-	gock.New("http://rpc1-dispute.localhost").
-		Post("/").
-		Times(1).
-		Reply(200).
-		SetHeader("Content-Type", "application/json").
-		JSON(map[string]interface{}{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  "0x7a",
-		})
-
-	gock.New("http://rpc2-dispute.localhost").
-		Post("/").
-		Times(1).
-		Reply(503).
-		SetHeader("Content-Type", "application/json").
-		JSON(map[string]interface{}{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"error": map[string]interface{}{
-				"code":    -32603,
-				"message": "unknown server error",
-			},
-		})
-
-	gock.New("http://rpc2-dispute.localhost").
-		Post("/").
-		Times(1).
-		Reply(200).
-		SetHeader("Content-Type", "application/json").
-		JSON(map[string]interface{}{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  "0x7a",
-		})
-
-	// Make request
-	reqBytes, err := json.Marshal(map[string]interface{}{
-		"method": "eth_getBlockByNumber",
-		"params": []interface{}{"0x77777", false},
-	})
-	if err != nil {
-		require.NoError(t, err)
-	}
-
-	fakeReq := common.NewNormalizedRequest(reqBytes)
-	resp, err := ntw.Forward(ctx, fakeReq)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	assert.NoError(t, err)
-	assert.NotNil(t, resp)
-
-	expectedJrr, err := common.NewJsonRpcResponse(1, "0x7a", nil)
-	assert.NoError(t, err)
-	assert.NotNil(t, expectedJrr)
-
-	actualJrr, err := resp.JsonRpcResponse()
-	assert.NoError(t, err)
-	assert.NotNil(t, actualJrr)
-}
-
-func pointer[T any](v T) *T {
-	return &v
-}
-
-// setupTestNetworkWithConsensusPolicy creates a test network with consensus policy for integration testing
-func setupTestNetworkWithConsensusPolicy(t *testing.T, ctx context.Context, upstreams []*common.UpstreamConfig, consensusConfig *common.ConsensusPolicyConfig) *Network {
-	// Apply defaults to consensus policy
-	err := consensusConfig.SetDefaults()
-	require.NoError(t, err)
-
-	mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
-	vr := thirdparty.NewVendorsRegistry()
-	pr, err := thirdparty.NewProvidersRegistry(&log.Logger, vr, []*common.ProviderConfig{}, nil)
-	require.NoError(t, err)
-
-	ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
-		Connector: &common.ConnectorConfig{
-			Driver: "memory",
-			Memory: &common.MemoryConnectorConfig{
-				MaxItems:     100_000,
-				MaxTotalSize: "1MB",
-			},
-		},
-	})
-	require.NoError(t, err)
-
-	upsReg := upstream.NewUpstreamsRegistry(
-		ctx, &log.Logger, "prjA", upstreams,
-		ssr, nil, vr, pr, nil, mt, 1*time.Second,
-	)
-
-	ntw, err := NewNetwork(
-		ctx, &log.Logger, "prjA",
-		&common.NetworkConfig{
-			Architecture: common.ArchitectureEvm,
-			Evm: &common.EvmNetworkConfig{
-				ChainId: 123,
-			},
-			Failsafe: []*common.FailsafeConfig{
-				{
-					MatchMethod: "*",
-					Consensus:   consensusConfig,
-				},
-			},
-		},
-		nil, upsReg, mt,
-	)
-	require.NoError(t, err)
-
-	err = upsReg.Bootstrap(ctx)
-	require.NoError(t, err)
-	err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(123))
-	require.NoError(t, err)
-
-	return ntw
-}
-
-// TestConsensusAcceptMostCommonValidResultScenarios tests AcceptMostCommonValidResult behavior with various scenarios
-func TestConsensusAcceptMostCommonValidResultScenarios(t *testing.T) {
-	tests1 := []struct {
-		name           string
-		mockResponses  []map[string]interface{}
-		expectedError  bool
-		expectedResult string
-		description    string
-	}{
-		{
-			name: "most_common_wins_2v1",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xaaa",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xaaa", // Same as first
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xbbb", // Different
-				},
-			},
-			expectedError:  false,
-			expectedResult: `"0xaaa"`,
-			description:    "2 identical non-empty vs 1 different → returns most common",
-		},
-		{
-			name: "empty_array_errors_2v1",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{}, // Empty array
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{}, // Empty array
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xccc", // Non-empty
-				},
-			},
-			expectedError:  false,
-			expectedResult: `"0xccc"`,
-			description:    "2 empty (meets threshold) vs 1 non-empty (below threshold) → returns non-empty",
-		},
-		{
-			name: "mostly_errors_one_empty_array_under_threshold",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{}, // Empty array
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32603,
-						"message": "Internal Server Error",
-					},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32603,
-						"message": "Internal Server Error",
-					},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"error": map[string]interface{}{
-						"code":    -32603,
-						"message": "Internal Server Error",
-					},
-				},
-			},
-			expectedError: true,
-			description:   "1 empty (below threshold) vs 3 errors (above threshold) → returns error",
-		},
-		{
-			name: "tie_2v2_no_clear_winner",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xaaa",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xaaa",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xbbb",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xbbb",
-				},
-			},
-			expectedError: true,
-			description:   "2 of result A, 2 of result B → errors (no clear winner)",
-		},
-		{
-			name: "clear_majority_4v1",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x123",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x123",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x123",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x123",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x999",
-				},
-			},
-			expectedError:  false,
-			expectedResult: `"0x123"`,
-			description:    "4 identical vs 1 different → returns result with count 4",
-		},
-		{
-			name: "clear_majority_3v2_alt",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xabc",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xabc",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xabc",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xdef",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xdef",
-				},
-			},
-			expectedError:  false,
-			expectedResult: `"0xabc"`,
-			description:    "3 identical vs 2 different → returns result with count 3",
-		},
-		{
-			name: "all_different_no_consensus",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x111",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x222",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x333",
-				},
-			},
-			expectedError: true,
-			description:   "All different responses → errors (no clear winner)",
-		},
-		{
-			name: "many_empty_few_nonempty_high_threshold",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xdata",
-				},
-			},
-			expectedError:  true,
-			expectedResult: "",
-			description:    "4 empty (meets threshold) vs 1 non-empty (below threshold) → errors",
-		},
-		{
-			name: "only_empty_results_insufficient_participants",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-			},
-			expectedError: true,
-			description:   "Only empty results with participants (2) < threshold (3) → error instead of empty result",
-		},
-		{
-			name: "non_empty_results_insufficient_participants",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xwinner",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xwinner",
-				},
-			},
-			expectedError: true,
-			description:   "non-empty similar results with participants (2) < threshold (3) → returns the result",
-		},
-		{
-			name: "only_empty_results_meets_threshold",
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-			},
-			expectedError:  false,
-			expectedResult: `[]`,
-			description:    "Only empty results with participants (3) >= threshold (2) → accept empty result",
-		},
-		// {
-		// 	name: "non_empty_below_threshold_error",
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0xwinner",
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0xwinner",
-		// 		},
-		// 	},
-		// 	expectedError: true, // AcceptMostCommon respects threshold
-		// 	description:   "Non-empty with count=2 < threshold=5 → error (threshold not met)",
-		// },
-		// {
-		// 	name: "competing_nonempty_with_empty_results",
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  []interface{}{}, // Empty
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  []interface{}{}, // Empty
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  []interface{}{}, // Empty
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0xresult1", // Non-empty 1
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0xresult2", // Non-empty 2 (competing)
-		// 		},
-		// 	},
-		// 	expectedError:  true,
-		// 	expectedResult: "",
-		// 	description:    "3 empty (meets threshold) vs 2 competing non-empty (below threshold) -> dispute",
-		// },
-		// {
-		// 	name: "empty_meets_threshold_wins",
-		// 	mockResponses: []map[string]interface{}{
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  []interface{}{}, // Empty
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  []interface{}{}, // Empty
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  []interface{}{}, // Empty
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0xwinner", // Non-empty winner (2 votes)
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0xwinner", // Non-empty winner (2 votes)
-		// 		},
-		// 		{
-		// 			"jsonrpc": "2.0",
-		// 			"id":      1,
-		// 			"result":  "0xminority", // Non-empty minority (1 vote)
-		// 		},
-		// 	},
-		// 	expectedError:  true,
-		// 	expectedResult: "",
-		// 	description:    "3 empty (meets threshold=3) vs 2 non-empty (below threshold) → errors",
-		// },
-	}
-
-	for _, tc := range tests1 {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer func() {
-				// Wait a moment before cleanup to avoid race with ongoing requests
-				time.Sleep(50 * time.Millisecond)
-				util.ResetGock()
-			}()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create upstreams with fully unique endpoints to guarantee deterministic mock matching
-			upstreams := make([]*common.UpstreamConfig, len(tc.mockResponses))
-			for i := range tc.mockResponses {
-				upstreamId := fmt.Sprintf("upstream%d", i+1)
-				// Use completely unique hostnames to eliminate any possibility of cross-contamination
-				uniqueHost := fmt.Sprintf("test-upstream-%d-%s.localhost", i+1, tc.name)
-				fullEndpoint := fmt.Sprintf("http://%s/rpc", uniqueHost)
-
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       upstreamId,
-					Endpoint: fullEndpoint,
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Create highly specific mocks with unique hosts - one mock per upstream
-			for i := range upstreams {
-				uniqueHost := fmt.Sprintf("test-upstream-%d-%s.localhost", i+1, tc.name)
-
-				gock.New(fmt.Sprintf("http://%s", uniqueHost)).
-					Post("/rpc").
-					Times(1).
-					Reply(200).
-					JSON(tc.mockResponses[i])
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network with AcceptMostCommonValidResult for disputes
-			// For tie test, we need higher threshold to force dispute
-			threshold := 2
-			if tc.name == "tie_2v2_no_clear_winner" {
-				threshold = 3 // This ensures neither result meets threshold
-			}
-			if tc.name == "only_empty_results_insufficient_participants" {
-				threshold = 3 // Set threshold higher than participants (2) to test insufficient participants scenario
-			}
-			if tc.name == "only_empty_results_meets_threshold" {
-				threshold = 2 // Set threshold lower than participants (3) to test sufficient participants scenario
-			}
-			if tc.name == "non_empty_below_threshold_error" {
-				threshold = 5 // Set threshold much higher than count (2) to test threshold requirement
-			}
-			if tc.name == "empty_meets_threshold_wins" {
-				threshold = 3 // Set threshold=3 so empty (3 votes) meets but non-empty (2 votes) doesn't
-			}
-			if tc.name == "clear_majority_3v2_alt" {
-				threshold = 3 // Set threshold=3 so only majority (3 votes) meets threshold, minority (2 votes) doesn't
-			}
-
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:    len(upstreams),
-				AgreementThreshold: threshold,
-				DisputeBehavior:    common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-				PunishMisbehavior:  &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				if err == nil && resp != nil {
-					jrr, _ := resp.JsonRpcResponse()
-					if jrr != nil {
-						t.Errorf("Expected error but got result: %s", string(jrr.Result))
-					}
-				}
-				assert.Error(t, err)
-				assert.True(t, common.HasErrorCode(err, common.ErrCodeConsensusDispute, common.ErrCodeConsensusLowParticipants))
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedResult, string(jrr.Result))
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-
-	// tests2 := []struct {
-	// 	name           string
-	// 	mockResponses  []map[string]interface{}
-	// 	hasErrors      bool
-	// 	expectedError  bool
-	// 	expectedResult interface{} // Can be string or predicate function
-	// 	description    string
-	// }{
-	// 	{
-	// 		name: "all_empty_returns_consensus",
-	// 		mockResponses: []map[string]interface{}{
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  []interface{}{},
-	// 			},
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  []interface{}{},
-	// 			},
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  []interface{}{},
-	// 			},
-	// 		},
-	// 		expectedError:  false,
-	// 		expectedResult: `[]`,
-	// 		description:    "All empty arrays → returns empty (primary consensus, not dispute)",
-	// 	},
-	// 	{
-	// 		name: "all_different_nonempty_returns_any",
-	// 		mockResponses: []map[string]interface{}{
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  "0xfirst",
-	// 			},
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  "0xsecond",
-	// 			},
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  "0xthird",
-	// 			},
-	// 		},
-	// 		expectedError: true,
-	// 		expectedResult: "",
-	// 		description: "All different non-empty → errors",
-	// 	},
-	// 	{
-	// 		name: "errors_and_empty_priority_ordering",
-	// 		mockResponses: []map[string]interface{}{
-	// 			nil, // Will trigger error
-	// 			nil, // Will trigger error
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  []interface{}{},
-	// 			},
-	// 		},
-	// 		hasErrors:      true,
-	// 		expectedError:  false, // With AcceptMostCommonValidResult, empty wins over errors
-	// 		expectedResult: `[]`,
-	// 		description:    "2 errors vs 1 empty → empty wins despite lower count (priority: empty > errors)",
-	// 	},
-	// 	{
-	// 		name: "errors_empty_nonempty_prefers_nonempty",
-	// 		mockResponses: []map[string]interface{}{
-	// 			nil, // Will trigger error
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  []interface{}{},
-	// 			},
-	// 			{
-	// 				"jsonrpc": "2.0",
-	// 				"id":      1,
-	// 				"result":  "0xbest",
-	// 			},
-	// 		},
-	// 		hasErrors:      true,
-	// 		expectedError:  false, // AcceptMostCommonValidResult should select non-empty result
-	// 		expectedResult: `"0xbest"`,
-	// 		description:    "1 error, 1 empty, 1 non-empty → prefer non-empty",
-	// 	},
-	// }
-
-	// for _, tc := range tests2 {
-	// 	t.Run(tc.name, func(t *testing.T) {
-	// 		util.ResetGock()
-	// 		defer util.ResetGock()
-	// 		util.SetupMocksForEvmStatePoller()
-
-	// 		// Create upstreams
-	// 		upstreams := make([]*common.UpstreamConfig, len(tc.mockResponses))
-	// 		for i := range tc.mockResponses {
-	// 			upstreams[i] = &common.UpstreamConfig{
-	// 				Id:       fmt.Sprintf("upstream%d", i+1),
-	// 				Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-	// 				Type:     common.UpstreamTypeEvm,
-	// 				Evm: &common.EvmUpstreamConfig{
-	// 					ChainId: 123,
-	// 				},
-	// 			}
-	// 		}
-
-	// 		// Mock responses
-	// 		for i, upstream := range upstreams {
-	// 			if tc.hasErrors && tc.mockResponses[i] == nil {
-	// 				// Simulate JSON-RPC error response
-	// 				gock.New(upstream.Endpoint).
-	// 					Post("").
-	// 					Times(1).
-	// 					Reply(200). // JSON-RPC errors use 200 status
-	// 					JSON(map[string]interface{}{
-	// 						"jsonrpc": "2.0",
-	// 						"id":      1,
-	// 						"error": map[string]interface{}{
-	// 							"code":    -32603,
-	// 							"message": "Internal Server Error",
-	// 						},
-	// 					})
-	// 			} else {
-	// 				gock.New(upstream.Endpoint).
-	// 					Post("").
-	// 					Times(1).
-	// 					Reply(200).
-	// 					JSON(tc.mockResponses[i])
-	// 			}
-	// 		}
-
-	// 		ctx, cancel := context.WithCancel(context.Background())
-	// 		defer cancel()
-
-	// 		// Create network with AcceptMostCommonValidResult for disputes
-	// 		network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-	// 			MaxParticipants:    len(upstreams),
-	// 			AgreementThreshold: 2,
-	// 			DisputeBehavior:    common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-	// 			PunishMisbehavior:  &common.PunishMisbehaviorConfig{},
-	// 		})
-
-	// 		// Make request
-	// 		req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-	// 		resp, err := network.Forward(ctx, req)
-
-	// 		if tc.expectedError {
-	// 			assert.Error(t, err)
-	// 		} else {
-	// 			require.NoError(t, err)
-	// 			require.NotNil(t, resp)
-
-	// 			jrr, err := resp.JsonRpcResponse()
-	// 			require.NoError(t, err)
-	// 			result := string(jrr.Result)
-
-	// 			// Handle both string and function expectations
-	// 			switch expected := tc.expectedResult.(type) {
-	// 			case string:
-	// 				assert.Equal(t, expected, result)
-	// 			case func(string) bool:
-	// 				assert.True(t, expected(result), "Result %s did not match predicate", result)
-	// 			}
-	// 		}
-
-	// 		t.Logf("%s: %s", tc.name, tc.description)
-	// 	})
-	// }
-}
-
-// TestConsensusDisputeBehaviorComparison tests different dispute behaviors with the same scenario
-func TestConsensusDisputeBehaviorComparison(t *testing.T) {
-	// Same dispute scenario: 3 different responses (no consensus possible)
-	disputeResponses := []map[string]interface{}{
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  "0xaaa",
-		},
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  "0xbbb",
-		},
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  []interface{}{}, // Empty array
-		},
-	}
-
-	tests := []struct {
-		name            string
-		disputeBehavior common.ConsensusDisputeBehavior
-		expectedError   bool
-		expectedResult  interface{} // Can be string or predicate function
-		description     string
-	}{
-		{
-			name:            "return_error_behavior",
-			disputeBehavior: common.ConsensusDisputeBehaviorReturnError,
-			expectedError:   true,
-			description:     "ReturnError → always returns error on dispute",
-		},
-		{
-			name:            "accept_most_common_no_winner",
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   true,
-			description:     "AcceptMostCommon → errors when no clear winner (all count 1)",
-		},
-		{
-			name:            "accept_any_valid_prefers_nonempty",
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false,
-			expectedResult: func(result string) bool {
-				// Should return one of the non-empty results
-				return result == `"0xaaa"` || result == `"0xbbb"`
-			},
-			description: "AcceptAnyValid → returns any valid result (prefers non-empty)",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create 3 upstreams
-			upstreams := make([]*common.UpstreamConfig, 3)
-			for i := range upstreams {
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       fmt.Sprintf("upstream%d", i+1),
-					Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Mock the same dispute responses for all tests
-			for i, upstream := range upstreams {
-				gock.New(upstream.Endpoint).
-					Post("").
-					Times(1).
-					Reply(200).
-					JSON(disputeResponses[i])
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network with specific dispute behavior
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:    3,
-				AgreementThreshold: 2,
-				DisputeBehavior:    tc.disputeBehavior,
-				PunishMisbehavior:  &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-				assert.True(t, common.HasErrorCode(err, common.ErrCodeConsensusDispute))
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				result := string(jrr.Result)
-
-				// Handle both string and function expectations
-				switch expected := tc.expectedResult.(type) {
-				case string:
-					assert.Equal(t, expected, result)
-				case func(string) bool:
-					assert.True(t, expected(result), "Result %s did not match predicate", result)
-				}
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-// TestConsensusLowParticipantsBehaviorComparison tests different low participants behaviors
-func TestConsensusLowParticipantsBehaviorComparison(t *testing.T) {
-	// Low participants scenario: Required 4, but only 2 respond
-	lowParticipantResponses := []map[string]interface{}{
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  "0xresponse1",
-		},
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  []interface{}{}, // Empty array
-		},
-	}
-
-	tests := []struct {
-		name                    string
-		lowParticipantsBehavior common.ConsensusLowParticipantsBehavior
-		expectedError           bool
-		expectedErrorCode       *common.ErrorCode
-		expectedResult          interface{} // Can be string or predicate function
-		description             string
-	}{
-		{
-			name:                    "return_error_behavior",
-			lowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
-			expectedError:           true,
-			expectedErrorCode:       pointer(common.ErrCodeConsensusLowParticipants),
-			description:             "ReturnError → returns error on low participants",
-		},
-		{
-			name:                    "accept_most_common_threshold_not_met",
-			lowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-			expectedError:           true, // No result meets threshold (all have count 1, threshold is 4)
-			expectedErrorCode:       pointer(common.ErrCodeConsensusLowParticipants),
-			description:             "AcceptMostCommon with 1 non-empty, 1 empty, 2 errors → error (no result meets threshold=4)",
-		},
-		{
-			name:                    "accept_any_valid_prefers_nonempty",
-			lowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-			expectedError:           false,
-			expectedResult:          `"0xresponse1"`, // Should prefer non-empty
-			description:             "AcceptAnyValid → returns non-empty result",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create 4 upstreams (but only 2 will respond)
-			upstreams := make([]*common.UpstreamConfig, 4)
-			for i := range upstreams {
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       fmt.Sprintf("upstream%d", i+1),
-					Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Mock only 2 responses (simulating low participants)
-			for i := 0; i < 2; i++ {
-				gock.New(upstreams[i].Endpoint).
-					Post("").
-					Times(1).
-					Reply(200).
-					JSON(lowParticipantResponses[i])
-			}
-
-			// Other upstreams timeout
-			for i := 2; i < 4; i++ {
-				gock.New(upstreams[i].Endpoint).
-					Post("").
-					Times(1).
-					Reply(500).
-					BodyString("timeout")
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network with specific low participants behavior
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:         4, // Require 4 but only 2 will respond
-				AgreementThreshold:      3, // Higher than participant count (2) to trigger low participants
-				LowParticipantsBehavior: tc.lowParticipantsBehavior,
-				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-				PunishMisbehavior:       &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-				if tc.expectedErrorCode != nil {
-					assert.True(t, common.HasErrorCode(err, *tc.expectedErrorCode))
-				}
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				result := string(jrr.Result)
-
-				// Handle both string and function expectations
-				switch expected := tc.expectedResult.(type) {
-				case string:
-					assert.Equal(t, expected, result)
-				case func(string) bool:
-					assert.True(t, expected(result), "Result %s did not match predicate", result)
-				}
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-// TestConsensusThresholdEdgeCases tests edge cases around agreement threshold
-func TestConsensusThresholdEdgeCases(t *testing.T) {
-	tests := []struct {
-		name               string
-		maxParticipants    int
-		agreementThreshold int
-		mockResponses      []map[string]interface{}
-		expectedError      bool
-		expectedErrorCode  *common.ErrorCode
-		expectedResult     string
-		description        string
-	}{
-		{
-			name:               "threshold_1_single_response",
-			maxParticipants:    1,
-			agreementThreshold: 1,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xsingle",
-				},
-			},
-			expectedError:  false,
-			expectedResult: `"0xsingle"`,
-			description:    "Threshold 1, 1 response → consensus achieved",
-		},
-		{
-			name:               "threshold_2_single_response_dispute",
-			maxParticipants:    2,
-			agreementThreshold: 2,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xonly",
-				},
-				nil, // Second upstream fails
-			},
-			expectedError:     true,
-			expectedErrorCode: pointer(common.ErrCodeConsensusLowParticipants),
-			description:       "Threshold 2, 1 response → low participants",
-		},
-		{
-			name:               "threshold_2_two_identical",
-			maxParticipants:    2,
-			agreementThreshold: 2,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xmatching",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xmatching",
-				},
-			},
-			expectedError:  false,
-			expectedResult: `"0xmatching"`,
-			description:    "Threshold 2, 2 identical → consensus achieved",
-		},
-		{
-			name:               "threshold_3_two_identical_dispute",
-			maxParticipants:    3,
-			agreementThreshold: 3,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xpair",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xpair",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xdifferent",
-				},
-			},
-			expectedError: true,
-			// Don't check specific error code - could be dispute or other depending on logic
-			description: "Threshold 3, 2 identical → error (below threshold)",
-		},
-		{
-			name:               "threshold_2_with_empty_consensus",
-			maxParticipants:    3,
-			agreementThreshold: 2,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xnonEmpty",
-				},
-			},
-			expectedError:     true,
-			expectedErrorCode: pointer(common.ErrCodeConsensusDispute),
-			description:       "Threshold 2, 2 empty arrays vs 1 non-empty → dispute (non-empty preference)",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create upstreams
-			upstreams := make([]*common.UpstreamConfig, tc.maxParticipants)
-			for i := range upstreams {
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       fmt.Sprintf("upstream%d", i+1),
-					Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Mock responses
-			for i, upstream := range upstreams {
-				if i < len(tc.mockResponses) && tc.mockResponses[i] != nil {
-					gock.New(upstream.Endpoint).
-						Post("").
-						Times(1).
-						Reply(200).
-						JSON(tc.mockResponses[i])
-				} else {
-					// Simulate failure
-					gock.New(upstream.Endpoint).
-						Post("").
-						Times(1).
-						Reply(500).
-						BodyString("error")
-				}
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:         tc.maxParticipants,
-				AgreementThreshold:      tc.agreementThreshold,
-				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
-				PunishMisbehavior:       &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-				if tc.expectedErrorCode != nil {
-					assert.True(t, common.HasErrorCode(err, *tc.expectedErrorCode))
-				}
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedResult, string(jrr.Result))
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-// TestConsensusEmptyNonEmptyPreferenceBehaviors tests empty/non-empty preference with different behaviors
-func TestConsensusEmptyNonEmptyPreferenceBehaviors(t *testing.T) {
-	// Same scenario: 3 empty, 1 non-empty with threshold 2
-	testResponses := []map[string]interface{}{
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  []interface{}{},
-		},
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  []interface{}{},
-		},
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  []interface{}{},
-		},
-		{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"result":  "0xvaluable",
-		},
-	}
-
-	tests := []struct {
-		name            string
-		disputeBehavior common.ConsensusDisputeBehavior
-		expectedError   bool
-		expectedResult  string
-		description     string
-	}{
-		{
-			name:            "accept_most_common_respects_threshold",
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false,
-			expectedResult:  `[]`, // Empty wins: count=3 meets threshold, non-empty count=1 doesn't
-			description:     "AcceptMostCommon → empty wins (3 meets threshold vs 1 non-empty below threshold)",
-		},
-		{
-			name:            "accept_any_valid_returns_nonempty",
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false,
-			expectedResult:  `"0xvaluable"`,
-			description:     "AcceptAnyValid → returns non-empty (ignores threshold)",
-		},
-		{
-			name:            "return_error_always_errors",
-			disputeBehavior: common.ConsensusDisputeBehaviorReturnError,
-			expectedError:   true,
-			description:     "ReturnError → always errors on dispute",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create 4 upstreams
-			upstreams := make([]*common.UpstreamConfig, 4)
-			for i := range upstreams {
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       fmt.Sprintf("upstream%d", i+1),
-					Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Mock responses
-			for i, upstream := range upstreams {
-				gock.New(upstream.Endpoint).
-					Post("").
-					Times(1).
-					Reply(200).
-					JSON(testResponses[i])
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:    4,
-				AgreementThreshold: 2,
-				DisputeBehavior:    tc.disputeBehavior,
-				PunishMisbehavior:  &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedResult, string(jrr.Result))
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-// TestConsensusComplexRealWorldScenarios tests complex real-world consensus scenarios
-func TestConsensusComplexRealWorldScenarios(t *testing.T) {
-	tests := []struct {
-		name                    string
-		maxParticipants         int
-		agreementThreshold      int
-		disputeBehavior         common.ConsensusDisputeBehavior
-		lowParticipantsBehavior common.ConsensusLowParticipantsBehavior
-		mockResponses           []map[string]interface{}
-		responseDelays          []time.Duration // Delays for each response
-		expectedError           bool
-		expectedResult          interface{} // Can be string or predicate function
-		description             string
-	}{
-		{
-			name:                    "low_participants_accept_any_mixed",
-			maxParticipants:         5,
-			agreementThreshold:      3,
-			disputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-			lowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{}, // Empty
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xdata", // Non-empty
-				},
-				nil, // Error
-				nil, // Error
-				nil, // Error
-			},
-			expectedError:  false,
-			expectedResult: `"0xdata"`, // AcceptAnyValid prefers non-empty
-			description:    "Low participants + AcceptAnyValid → prefers non-empty over empty",
-		},
-		{
-			name:                    "dispute_threshold_not_met",
-			maxParticipants:         4,
-			agreementThreshold:      3,
-			disputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			lowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xvalue",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xvalue",
-				},
-			},
-			expectedError: true, // No result meets threshold=3 (both have count=2)
-			description:   "Dispute with 2 empty vs 2 non-empty → error (no result meets threshold=3)",
-		},
-		{
-			name:                    "short_circuit_prevention_late_nonempty",
-			maxParticipants:         3,
-			agreementThreshold:      2,
-			disputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			lowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xlate", // Late non-empty response
-				},
-			},
-			responseDelays: []time.Duration{
-				0,
-				0,
-				200 * time.Millisecond, // Delayed response
-			},
-			expectedError:  false,
-			expectedResult: `"0xlate"`, // Should wait and prefer non-empty
-			description:    "Short-circuit prevention waits for late non-empty response",
-		},
-		{
-			name:                    "complex_mixed_scenario",
-			maxParticipants:         6,
-			agreementThreshold:      3,
-			disputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			lowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-			mockResponses: []map[string]interface{}{
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xcommon",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xcommon",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xcommon",
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  []interface{}{},
-				},
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0xdifferent",
-				},
-				nil, // Error
-			},
-			expectedError:  false,
-			expectedResult: `"0xcommon"`, // Clear winner with count 3
-			description:    "Complex scenario with clear consensus despite mixed responses",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create upstreams
-			upstreams := make([]*common.UpstreamConfig, tc.maxParticipants)
-			for i := range upstreams {
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       fmt.Sprintf("upstream%d", i+1),
-					Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Mock responses with delays
-			for i, upstream := range upstreams {
-				if i < len(tc.mockResponses) {
-					if tc.mockResponses[i] == nil {
-						// Error response
-						gock.New(upstream.Endpoint).
-							Post("").
-							Times(1).
-							Reply(500).
-							BodyString("error")
-					} else {
-						mock := gock.New(upstream.Endpoint).
-							Post("").
-							Times(1).
-							Reply(200)
-
-						// Add delay if specified
-						if tc.responseDelays != nil && i < len(tc.responseDelays) && tc.responseDelays[i] > 0 {
-							mock = mock.Delay(tc.responseDelays[i])
-						}
-
-						mock.JSON(tc.mockResponses[i])
-					}
-				}
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:         tc.maxParticipants,
-				AgreementThreshold:      tc.agreementThreshold,
-				DisputeBehavior:         tc.disputeBehavior,
-				LowParticipantsBehavior: tc.lowParticipantsBehavior,
-				PunishMisbehavior:       &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				result := string(jrr.Result)
-
-				// Handle both string and function expectations
-				switch expected := tc.expectedResult.(type) {
-				case string:
-					assert.Equal(t, expected, result)
-				case func(string) bool:
-					assert.True(t, expected(result), "Result %s did not match predicate", result)
-				}
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-// TestConsensusMixedErrorScenarios tests consensus with various error types
-func TestConsensusMixedErrorScenarios(t *testing.T) {
-	tests := []struct {
-		name           string
-		mockResponses  []interface{} // Can be map for success or error config
-		expectedError  bool
-		expectedResult interface{} // Can be string or error matcher function
-		description    string
-	}{
-		{
-			name: "consensus_on_same_execution_error",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-			},
-			expectedError: true,
-			expectedResult: func(err error) bool {
-				if err == nil {
-					return false
-				}
-				// Should return the agreed-upon execution error
-				return strings.Contains(err.Error(), "execution reverted") &&
-					strings.Contains(err.Error(), "insufficient balance")
-			},
-			description: "3 identical execution errors → consensus on error",
-		},
-		{
-			name: "different_execution_errors_return_first_error",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: reason A",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: reason B",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: reason C",
-					},
-				},
-			},
-			expectedError: true,
-			expectedResult: func(err error) bool {
-				if err == nil {
-					return false
-				}
-				// With ReturnError dispute behavior, it returns one of the execution errors
-				return strings.Contains(err.Error(), "execution reverted") &&
-					(strings.Contains(err.Error(), "reason A") ||
-						strings.Contains(err.Error(), "reason B") ||
-						strings.Contains(err.Error(), "reason C"))
-			},
-			description: "3 different execution errors → returns first error (ReturnError behavior)",
-		},
-		{
-			name: "success_wins_over_fewer_execution_errors",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    3,
-						"message": "execution reverted: token paused",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    3,
-						"message": "execution reverted: token paused",
-					},
-				},
-				map[string]interface{}{
-					"result": "0x123", // Success response
-				},
-			},
-			expectedError:  false, // Success response wins due to non-empty preference
-			expectedResult: `"0x123"`,
-			description:    "2 same execution errors, 1 success → success wins (non-empty preference)",
-		},
-		{
-			name: "success_wins_over_execution_errors_non_empty_preference",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000, // Use standard execution error code
-						"message": "execution reverted: out of gas",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: out of gas",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: out of gas",
-					},
-				},
-				map[string]interface{}{
-					"result": "0x456", // Success response wins due to non-empty preference
-				},
-			},
-			expectedError:  false, // Success response wins due to non-empty preference
-			expectedResult: `"0x456"`,
-			description:    "3 same execution errors, 1 success → success wins (non-empty preference)",
-		},
-		{
-			name: "mix_network_and_execution_errors",
-			mockResponses: []interface{}{
-				"network_error",
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    3,
-						"message": "execution reverted: out of gas",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    3,
-						"message": "execution reverted: out of gas",
-					},
-				},
-			},
-			expectedError: true,
-			expectedResult: func(err error) bool {
-				// Execution errors should form consensus
-				return strings.Contains(err.Error(), "execution reverted") &&
-					strings.Contains(err.Error(), "out of gas")
-			},
-			description: "1 network error, 2 same execution errors → consensus on execution error",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create upstreams with unique endpoints from the start to avoid race conditions
-			upstreams := make([]*common.UpstreamConfig, len(tc.mockResponses))
-			for i := range upstreams {
-				upstreamId := fmt.Sprintf("upstream%d", i+1)
-				baseUrl := fmt.Sprintf("http://upstream%d.localhost", i+1)
-				uniquePath := fmt.Sprintf("/%s", upstreamId)
-				fullEndpoint := baseUrl + uniquePath
-
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       upstreamId,
-					Endpoint: fullEndpoint, // Use unique endpoint from creation
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Mock responses with unique paths
-			for i, upstream := range upstreams {
-				upstreamId := upstream.Id
-				baseUrl := fmt.Sprintf("http://upstream%d.localhost", i+1)
-				uniquePath := fmt.Sprintf("/%s", upstreamId)
-
-				switch resp := tc.mockResponses[i].(type) {
-				case string:
-					if resp == "network_error" {
-						// Simulate network error
-						gock.New(baseUrl).
-							Post(uniquePath).
-							Times(1). // Use Times(1) with unique paths for deterministic behavior
-							ReplyError(fmt.Errorf("network timeout"))
-					}
-				case map[string]interface{}:
-					// Normal JSON-RPC response (success or error)
-					response := map[string]interface{}{
-						"jsonrpc": "2.0",
-						"id":      1,
-					}
-					// Copy fields from test response
-					for k, v := range resp {
-						response[k] = v
-					}
-					gock.New(baseUrl).
-						Post(uniquePath).
-						Times(1). // Use Times(1) with unique paths for deterministic behavior
-						Reply(200).
-						JSON(response)
-				}
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network with consensus
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:    len(upstreams),
-				AgreementThreshold: 2,
-				DisputeBehavior:    common.ConsensusDisputeBehaviorReturnError,
-				PunishMisbehavior:  &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_call","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-				// Check error matches expectation
-				switch expected := tc.expectedResult.(type) {
-				case func(error) bool:
-					assert.True(t, expected(err), "Error %v did not match predicate", err)
-				}
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedResult, string(jrr.Result))
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-// TestConsensusDisputeBehaviorWithMinorityNonEmpty tests specific scenarios where non-empty responses are minority
-func TestConsensusDisputeBehaviorWithMinorityNonEmpty(t *testing.T) {
-	tests := []struct {
-		name            string
-		mockResponses   []interface{}
-		disputeBehavior common.ConsensusDisputeBehavior
-		expectedError   bool
-		expectedResult  interface{} // Can be string or error matcher function
-		description     string
-	}{
-
-		{
-			name: "accept_any_valid_prefers_single_success_over_execution_errors",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-				map[string]interface{}{
-					"result": "0x456",
-				},
-			},
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false,
-			expectedResult:  `"0x456"`,
-			description:     "1 valid non-empty response, 2 execution errors + acceptAnyValidResult → returns non-empty",
-		},
-		{
-			name: "accept_most_common_returns_success_vs_execution_errors",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-				map[string]interface{}{
-					"result": "0x456",
-				},
-			},
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false, // Success response wins due to non-empty preference
-			expectedResult:  `"0x456"`,
-			description:     "1 valid non-empty response, 2 execution errors + acceptMostCommonValidResult → returns non-empty (non-empty preference)",
-		},
-		{
-			name: "accept_most_common_errors_with_different_non_empty_responses",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"result": "0x111",
-				},
-				map[string]interface{}{
-					"result": "0x222",
-				},
-				map[string]interface{}{
-					"result": "0x333",
-				},
-			},
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   true, // All different non-empty responses, no clear winner
-			expectedResult: func(err error) bool {
-				if err == nil {
-					return false
-				}
-				// Should be a dispute error because not enough participants
-				return common.HasErrorCode(err, common.ErrCodeConsensusDispute)
-			},
-			description: "3 different non-empty responses + acceptMostCommonValidResult → error (no clear winner)",
-		},
-		{
-			name: "accept_any_valid_rejects_empty_with_missing_data_errors",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"result": []interface{}{}, // Empty array response
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32602, // Missing data error
-						"message": "missing trie node",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32602, // Missing data error
-						"message": "missing trie node",
-					},
-				},
-			},
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false, // AcceptMostCommonValidResult accepts the empty array as a valid result
-			expectedResult:  "[]",  // Empty array result
-			description:     "1 empty response, 2 missing data errors + acceptAnyValidResult → empty array accepted",
-		},
-		{
-			name: "accept_most_common_empty_consensus_only_empty_responses",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"result": []interface{}{}, // Empty array response
-				},
-				map[string]interface{}{
-					"result": []interface{}{}, // Empty array response
-				},
-			},
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false, // Empty responses achieve consensus (2 empty, threshold 2)
-			expectedResult:  `[]`,
-			description:     "2 empty responses + acceptMostCommonValidResult → returns empty (consensus)",
-		},
-		{
-			name: "accept_any_valid_empty_consensus_only_empty_responses",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"result": []interface{}{}, // Empty array response
-				},
-				map[string]interface{}{
-					"result": []interface{}{}, // Empty array response
-				},
-			},
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   false, // Empty responses achieve consensus (2 empty, threshold 2)
-			expectedResult:  `[]`,
-			description:     "2 empty responses + acceptAnyValidResult → returns empty (consensus)",
-		},
-		{
-			name: "accept_any_valid_returns_consensus_error_correctly",
-			mockResponses: []interface{}{
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-				map[string]interface{}{
-					"error": map[string]interface{}{
-						"code":    -32000,
-						"message": "execution reverted: insufficient balance",
-					},
-				},
-			},
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-			expectedError:   true, // Should return the consensus-valid error
-			expectedResult: func(err error) bool {
-				if err == nil {
-					return false
-				}
-				// Should be the execution error, not a dispute error
-				return strings.Contains(err.Error(), "execution reverted: insufficient balance")
-			},
-			description: "2 consensus-valid errors + acceptAnyValidResult → returns consensus error (not dispute)",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create upstreams
-			upstreams := make([]*common.UpstreamConfig, len(tc.mockResponses))
-			for i := range upstreams {
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       fmt.Sprintf("upstream%d", i+1),
-					Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-			}
-
-			// Mock responses
-			for i, upstream := range upstreams {
-				switch resp := tc.mockResponses[i].(type) {
-				case string:
-					if resp == "network_error" {
-						// Simulate network error
-						gock.New(upstream.Endpoint).
-							Post("").
-							Times(1).
-							ReplyError(fmt.Errorf("network timeout"))
-					}
-				case map[string]interface{}:
-					// Normal JSON-RPC response (success or error)
-					response := map[string]interface{}{
-						"jsonrpc": "2.0",
-						"id":      1,
-					}
-					// Copy fields from test response
-					for k, v := range resp {
-						response[k] = v
-					}
-					gock.New(upstream.Endpoint).
-						Post("").
-						Times(1).
-						Reply(200).
-						JSON(response)
-				}
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network with the specific dispute behavior
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:    len(upstreams),
-				AgreementThreshold: 2,
-				DisputeBehavior:    tc.disputeBehavior,
-				PunishMisbehavior:  &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_call","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-				// Check error matches expectation
-				switch expected := tc.expectedResult.(type) {
-				case func(error) bool:
-					assert.True(t, expected(err), "Error %v did not match predicate", err)
-				}
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedResult, string(jrr.Result))
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-// TestConsensusParticipantRequirements tests basic participant requirement scenarios
-// Note: Complex timeout scenarios are covered by existing integration tests
-func TestConsensusParticipantRequirements(t *testing.T) {
-	tests := []struct {
-		name               string
-		upstreams          int
-		maxParticipants    int
-		agreementThreshold int
-		mockResponses      []string
-		expectedError      bool
-		expectedErrorCode  *common.ErrorCode
-		expectedResult     string
-		description        string
-	}{
-		{
-			name:               "normal_consensus_achieved",
-			upstreams:          3,
-			maxParticipants:    3,
-			agreementThreshold: 2,
-			mockResponses:      []string{"0xaaa", "0xaaa", "0xbbb"},
-			expectedError:      false,
-			expectedResult:     `"0xaaa"`,
-			description:        "Normal consensus: 3 participants, 2 agree → consensus",
-		},
-		{
-			name:               "low_threshold_single_response",
-			upstreams:          3,
-			maxParticipants:    1,
-			agreementThreshold: 1,
-			mockResponses:      []string{"0xsame", "0xsame", "0xsame"},
-			expectedError:      false,
-			expectedResult:     `"0xsame"`, // All same response, threshold 1
-			description:        "Low threshold: 1 required, threshold 1 → consensus achieved",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create upstreams
-			upstreams := make([]*common.UpstreamConfig, tc.upstreams)
-			for i := 0; i < tc.upstreams; i++ {
-				upstreams[i] = &common.UpstreamConfig{
-					Id:       fmt.Sprintf("upstream%d", i+1),
-					Endpoint: fmt.Sprintf("http://upstream%d.localhost", i+1),
-					Type:     common.UpstreamTypeEvm,
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				}
-
-				// Mock successful responses for all upstreams
-				gock.New(upstreams[i].Endpoint).
-					Post("").
-					Times(1).
-					Reply(200).
-					JSON(map[string]interface{}{
-						"jsonrpc": "2.0",
-						"id":      1,
-						"result":  tc.mockResponses[i],
-					})
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			// Create network
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, &common.ConsensusPolicyConfig{
-				MaxParticipants:         tc.maxParticipants,
-				AgreementThreshold:      tc.agreementThreshold,
-				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
-				PunishMisbehavior:       &common.PunishMisbehaviorConfig{},
-			})
-
-			// Make request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_chainId","params":[],"id":1}`))
-			resp, err := network.Forward(ctx, req)
-
-			if tc.expectedError {
-				assert.Error(t, err)
-				if tc.expectedErrorCode != nil {
-					assert.True(t, common.HasErrorCode(err, *tc.expectedErrorCode),
-						"Expected error code %v but got: %v", *tc.expectedErrorCode, err)
-				}
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, resp)
-
-				jrr, err := resp.JsonRpcResponse()
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedResult, string(jrr.Result))
-			}
-
-			t.Logf("%s: %s", tc.name, tc.description)
-		})
-	}
-}
-
-func TestNetwork_ConsensusOnAgreedErrors(t *testing.T) {
-	// Helper to create JSON-RPC error response
-	createErrorResponse := func(code int, message string) map[string]interface{} {
-		return map[string]interface{}{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"error": map[string]interface{}{
-				"code":    code,
-				"message": message,
-			},
-		}
-	}
-
-	tests := []struct {
-		name              string
-		upstreams         []*common.UpstreamConfig
-		request           map[string]interface{}
-		mockResponses     []map[string]interface{}
-		maxParticipants   int
-		expectedError     bool
-		expectedErrorCode *int // Expected JSON-RPC error code
-		expectedErrorMsg  string
-	}{
-		{
-			name:            "consensus_on_invalid_params_error",
-			maxParticipants: 3,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_call",
-				"params": []interface{}{}, // Invalid params
-			},
-			mockResponses: []map[string]interface{}{
-				createErrorResponse(-32602, "Invalid params"),
-				createErrorResponse(-32602, "Invalid params"),
-				createErrorResponse(-32602, "Invalid params"),
-			},
-			expectedError:     true,
-			expectedErrorCode: pointer(-32602),
-			expectedErrorMsg:  "Invalid params",
-		},
-		{
-			name:            "consensus_on_method_not_found",
-			maxParticipants: 3,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_unknownMethod",
-				"params": []interface{}{},
-			},
-			mockResponses: []map[string]interface{}{
-				createErrorResponse(-32601, "Method not found"),
-				createErrorResponse(-32601, "Method not found"),
-				createErrorResponse(-32601, "Method not found"),
-			},
-			expectedError:     true,
-			expectedErrorCode: pointer(-32601),
-			expectedErrorMsg:  "Method not found",
-		},
-		{
-			name:            "consensus_on_missing_data",
-			maxParticipants: 3,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_getBlockByNumber",
-				"params": []interface{}{"0x999999999", false}, // Very old block
-			},
-			mockResponses: []map[string]interface{}{
-				createErrorResponse(-32014, "requested data is not available"),
-				createErrorResponse(-32014, "requested data is not available"),
-				createErrorResponse(-32014, "requested data is not available"),
-			},
-			expectedError:     true,
-			expectedErrorCode: pointer(-32014),
-			expectedErrorMsg:  "requested data is not available",
-		},
-		{
-			name:            "consensus_on_execution_reverted",
-			maxParticipants: 3,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_call",
-				"params": []interface{}{
-					map[string]interface{}{
-						"to":   "0x0000000000000000000000000000000000000000",
-						"data": "0x12345678",
-					},
-					"latest",
-				},
-			},
-			mockResponses: []map[string]interface{}{
-				createErrorResponse(3, "execution reverted"),
-				createErrorResponse(3, "execution reverted"),
-				createErrorResponse(3, "execution reverted"),
-			},
-			expectedError:     true,
-			expectedErrorCode: pointer(3),
-			expectedErrorMsg:  "execution reverted",
-		},
-		{
-			name:            "mixed_errors_returns_dispute",
-			maxParticipants: 3,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_call",
-				"params": []interface{}{},
-			},
-			mockResponses: []map[string]interface{}{
-				createErrorResponse(-32602, "Invalid params"),
-				createErrorResponse(-32601, "Method not found"),
-				createErrorResponse(-32000, "Internal error"),
-			},
-			expectedError:     true,
-			expectedErrorCode: nil, // Should get dispute error, not specific JSON-RPC error
-			expectedErrorMsg:  "not enough participants",
-		},
-		{
-			name:            "majority_agreement_on_error_with_one_success",
-			maxParticipants: 3,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "test1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc1-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc2-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-				{
-					Id:       "test3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://rpc3-agreed-error.localhost",
-					Evm: &common.EvmUpstreamConfig{
-						ChainId: 123,
-					},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_call",
-				"params": []interface{}{},
-			},
-			mockResponses: []map[string]interface{}{
-				createErrorResponse(-32602, "Invalid params"),
-				createErrorResponse(-32602, "Invalid params"),
-				{
-					"jsonrpc": "2.0",
-					"id":      1,
-					"result":  "0x1234",
-				},
-			},
-			expectedError:     true,
-			expectedErrorCode: pointer(-32602),
-			expectedErrorMsg:  "Invalid params",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-			// Don't check for pending mocks since we use Persist() for consensus scenarios
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer func() {
-				cancel()
-				time.Sleep(100 * time.Millisecond)
-			}()
-
-			// Setup network infrastructure
-			mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
-			vr := thirdparty.NewVendorsRegistry()
-			pr, err := thirdparty.NewProvidersRegistry(&log.Logger, vr, []*common.ProviderConfig{}, nil)
-			require.NoError(t, err)
-
-			ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
-				Connector: &common.ConnectorConfig{
-					Driver: "memory",
-					Memory: &common.MemoryConnectorConfig{
-						MaxItems:     100_000,
-						MaxTotalSize: "1MB",
-					},
-				},
-			})
-			require.NoError(t, err)
-
-			upsReg := upstream.NewUpstreamsRegistry(
-				ctx, &log.Logger, "prjA", tt.upstreams,
-				ssr, nil, vr, pr, nil, mt, 1*time.Second,
-			)
-
-			// Create network with consensus policy
-			ntw, err := NewNetwork(
-				ctx, &log.Logger, "prjA",
-				&common.NetworkConfig{
-					Architecture: common.ArchitectureEvm,
-					Evm: &common.EvmNetworkConfig{
-						ChainId: 123,
-					},
-					Failsafe: []*common.FailsafeConfig{
-						{
-							MatchMethod: "*",
-							Consensus: &common.ConsensusPolicyConfig{
-								MaxParticipants:         tt.maxParticipants,
-								AgreementThreshold:      2, // Majority agreement
-								DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-								LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
-							},
-						},
-					},
-				},
-				nil, upsReg, mt,
-			)
-			require.NoError(t, err)
-
-			err = upsReg.Bootstrap(ctx)
-			require.NoError(t, err)
-			err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(123))
-			require.NoError(t, err)
-
-			// Setup mock responses
-			for i, upstream := range tt.upstreams {
-				gock.New(upstream.Endpoint).
-					Post("/").
-					Persist().
-					Reply(200).
-					SetHeader("Content-Type", "application/json").
-					JSON(tt.mockResponses[i])
-			}
-
-			// Make request
-			reqBytes, err := json.Marshal(tt.request)
-			require.NoError(t, err)
-
-			fakeReq := common.NewNormalizedRequest(reqBytes)
-			resp, err := ntw.Forward(ctx, fakeReq)
-
-			// Verify results
-			if tt.expectedError {
-				assert.Error(t, err)
-				assert.Nil(t, resp)
-
-				if tt.expectedErrorCode != nil {
-					// Should get the specific JSON-RPC error
-					var jre *common.ErrJsonRpcExceptionInternal
-					if assert.True(t, errors.As(err, &jre), "expected JSON-RPC error but got %T: %v", err, err) {
-						assert.Equal(t, common.JsonRpcErrorNumber(*tt.expectedErrorCode), jre.NormalizedCode(),
-							"expected error code %d but got %d", *tt.expectedErrorCode, jre.NormalizedCode())
-						assert.Contains(t, jre.Message, tt.expectedErrorMsg)
-					}
-				} else {
-					// Should get consensus dispute error
-					assert.True(t, common.HasErrorCode(err, common.ErrCodeConsensusDispute, common.ErrCodeConsensusLowParticipants),
-						"expected consensus dispute error but got %v", err)
-					assert.Contains(t, err.Error(), tt.expectedErrorMsg)
-				}
-			} else {
-				assert.NoError(t, err)
-				assert.NotNil(t, resp)
-			}
-		})
-	}
-}
-
-func TestNetwork_ConsensusPolicy_LargerResponsePreference(t *testing.T) {
-	// Real-world inspired test data based on HyperEVM discrepancy example
-	// Good response (larger, more complete)
-	largerResponse := map[string]interface{}{
-		"jsonrpc": "2.0",
-		"id":      1,
-		"result": []interface{}{
-			map[string]interface{}{
-				"blockHash":         "0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7",
-				"blockNumber":       "0x9c0400",
-				"contractAddress":   nil,
-				"cumulativeGasUsed": "0x0",
-				"effectiveGasPrice": "0x0",
-				"from":              "0x0000000000000000000000000000000000000000",
-				"gasUsed":           "0x0",
-				"logs": []interface{}{
-					map[string]interface{}{
-						"address": "0x222222222264d28317734c14d2b8e7c1803ca8f6", // Transfer event
-						"topics": []interface{}{
-							"0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef",
-							"0x0000000000000000000000000000000000000000000000000000000000000000",
-							"0x0000000000000000000000007be8f48894d9ec0528ca70d9151cf2831c377be0",
-						},
-						"data":             "0x00000000000000000000000000000000000000000000021e19e0c9bab2400000",
-						"blockNumber":      "0x9c0400",
-						"transactionHash":  "0x3485d0657cc683514d4b8d2ef82881e5a77af11d1562ff3200614afe5a73efea",
-						"transactionIndex": "0x0",
-						"blockHash":        "0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7",
-						"logIndex":         "0x0",
-						"removed":          false,
-					},
-					map[string]interface{}{
-						"address": "0x7be8f48894d9ec0528ca70d9151cf2831c377be0",
-						"topics": []interface{}{
-							"0x2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d",
-						},
-						"data":             "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-						"blockNumber":      "0x9c0400",
-						"transactionHash":  "0x3485d0657cc683514d4b8d2ef82881e5a77af11d1562ff3200614afe5a73efea",
-						"transactionIndex": "0x0",
-						"blockHash":        "0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7",
-						"logIndex":         "0x1",
-						"removed":          false,
-					},
-				},
-				"logsBloom":        "0x00000000000000000000010000000000000000000000000020000000000100000000000000000000000000000000000000000000000000000000000000000000000002000000000000000008000000000000000000000000000000000000000010000000000000000000000000040000000000000000000000000010000000000000000000000000000000000000000000000000000000080000000000000001000000000000000000000000000000000100000000000000000000000000000000000002000000000080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000040000000000000800",
-				"status":           "0x1",
-				"to":               "0x7be8f48894d9ec0528ca70d9151cf2831c377be0",
-				"transactionHash":  "0x3485d0657cc683514d4b8d2ef82881e5a77af11d1562ff3200614afe5a73efea",
-				"transactionIndex": "0x0",
-				"type":             "0x2",
-			},
-		},
-	}
-
-	// Bad response (smaller, incomplete - missing logs and important fields)
-	smallerResponse := map[string]interface{}{
-		"jsonrpc": "2.0",
-		"id":      1,
-		"result": []interface{}{
-			map[string]interface{}{
-				"blockHash":         "0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7",
-				"blockNumber":       "0x9c0400",
-				"contractAddress":   nil,
-				"cumulativeGasUsed": "0x0",
-				"status":            "0x1",
-				"transactionHash":   "0x3485d0657cc683514d4b8d2ef82881e5a77af11d1562ff3200614afe5a73efea",
-				"transactionIndex":  "0x0",
-			},
-		},
-	}
-
-	tests := []struct {
-		name                     string
-		upstreams                []*common.UpstreamConfig
-		request                  map[string]interface{}
-		mockResponses            []map[string]interface{}
-		maxParticipants          int
-		agreementThreshold       int
-		preferNonEmpty           *bool
-		preferLargerResponses    *bool
-		expectedResponseSize     string // "larger" or "smaller"
-		expectedResponseContains string // Check for specific content
-	}{
-		{
-			name:            "larger_response_wins_with_equal_votes_when_both_meet_threshold",
-			maxParticipants: 4,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "good-vendor1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "good-vendor2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "bad-vendor1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "bad-vendor2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_getBlockReceipts",
-				"params": []interface{}{"0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7"},
-			},
-			mockResponses: []map[string]interface{}{
-				largerResponse, // 2 votes for larger, complete response
-				largerResponse,
-				smallerResponse, // 2 votes for smaller, incomplete response
-				smallerResponse,
-			},
-			agreementThreshold:       2,
-			preferLargerResponses:    pointer(true),
-			expectedResponseSize:     "larger",
-			expectedResponseContains: "0x222222222264d28317734c14d2b8e7c1803ca8f6", // Transfer event address
-		},
-		{
-			name:            "larger_response_wins_with_fewer_votes_when_both_meet_threshold",
-			maxParticipants: 5,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "good-vendor1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "good-vendor2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "bad-vendor1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "bad-vendor2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "bad-vendor3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor3.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_getBlockReceipts",
-				"params": []interface{}{"0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7"},
-			},
-			mockResponses: []map[string]interface{}{
-				largerResponse, // 2 votes for larger, complete response (meets threshold)
-				largerResponse,
-				smallerResponse, // 3 votes for smaller, incomplete response (meets threshold)
-				smallerResponse,
-				smallerResponse,
-			},
-			agreementThreshold:       2,
-			preferLargerResponses:    pointer(true),
-			expectedResponseSize:     "larger",                                     // Should win despite fewer votes because it's larger
-			expectedResponseContains: "0x222222222264d28317734c14d2b8e7c1803ca8f6", // Transfer event address
-		},
-		{
-			name:            "smaller_response_wins_with_more_votes_when_prefer_larger_disabled",
-			maxParticipants: 5,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "nirvana-hyperevm",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "hyperliquid-hyperlend",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "quicknode-hyperevm",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "hyperliquid-blockpi",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "hyperliquid-imperator",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor3.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_getBlockReceipts",
-				"params": []interface{}{"0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7"},
-			},
-			mockResponses: []map[string]interface{}{
-				largerResponse,
-				largerResponse,
-				smallerResponse,
-				smallerResponse,
-				smallerResponse,
-			},
-			agreementThreshold:    2,
-			preferLargerResponses: pointer(false), // Disable larger response preference
-			expectedResponseSize:  "smaller",
-		},
-		{
-			name:            "larger_response_only_wins_if_meets_threshold",
-			maxParticipants: 4,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "nirvana-hyperevm",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "quicknode-hyperevm",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "hyperliquid-blockpi",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "hyperliquid-imperator",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://bad-vendor3.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_getBlockReceipts",
-				"params": []interface{}{"0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7"},
-			},
-			mockResponses: []map[string]interface{}{
-				largerResponse,  // 1 vote for larger response (doesn't meet threshold of 2)
-				smallerResponse, // 3 votes for smaller response (meets threshold)
-				smallerResponse,
-				smallerResponse,
-			},
-			agreementThreshold:    2,
-			preferLargerResponses: pointer(true),
-			expectedResponseSize:  "smaller", // Smaller wins because larger doesn't meet threshold
-		},
-		{
-			name:            "equal_size_responses_fallback_to_count",
-			maxParticipants: 4,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "vendor1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "vendor2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "vendor3",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://vendor3.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "vendor4",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://vendor4.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_getBlockReceipts",
-				"params": []interface{}{"0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7"},
-			},
-			mockResponses: []map[string]interface{}{
-				largerResponse, // 3 votes for larger, complete response
-				largerResponse,
-				largerResponse,
-				smallerResponse, // 1 vote for smaller response
-			},
-			agreementThreshold:       2,
-			preferLargerResponses:    pointer(true),
-			expectedResponseSize:     "larger", // Should pick the larger response (more votes + larger)
-			expectedResponseContains: "0x222222222264d28317734c14d2b8e7c1803ca8f6",
-		},
-		{
-			name:            "larger_response_preferred_with_non_empty_preference_disabled",
-			maxParticipants: 4,
-			upstreams: []*common.UpstreamConfig{
-				{
-					Id:       "good-vendor1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "good-vendor2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://good-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "error-vendor1",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://error-vendor1.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-				{
-					Id:       "error-vendor2",
-					Type:     common.UpstreamTypeEvm,
-					Endpoint: "http://error-vendor2.localhost",
-					Evm:      &common.EvmUpstreamConfig{ChainId: 998},
-				},
-			},
-			request: map[string]interface{}{
-				"method": "eth_getBlockReceipts",
-				"params": []interface{}{"0x2033dc0b73dc048827289e5e9b33dc70d9b67c6ca8141400f2e917b14f1fa4e7"},
-			},
-			mockResponses: []map[string]interface{}{
-				largerResponse, // 2 votes for good response
-				largerResponse,
-				{"jsonrpc": "2.0", "id": 1, "error": map[string]interface{}{"code": -32603, "message": "execution reverted"}}, // 2 votes for error
-				{"jsonrpc": "2.0", "id": 1, "error": map[string]interface{}{"code": -32603, "message": "execution reverted"}},
-			},
-			agreementThreshold:       2,
-			preferNonEmpty:           pointer(false), // All response types compete equally
-			preferLargerResponses:    pointer(true),
-			expectedResponseSize:     "larger", // Should pick the larger successful response
-			expectedResponseContains: "0x222222222264d28317734c14d2b8e7c1803ca8f6",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Reset HTTP mocks
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer func() {
-				cancel()
-				time.Sleep(100 * time.Millisecond)
-			}()
-
-			// Setup network infrastructure
-			mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
-			vr := thirdparty.NewVendorsRegistry()
-			pr, err := thirdparty.NewProvidersRegistry(&log.Logger, vr, []*common.ProviderConfig{}, nil)
-			require.NoError(t, err)
-
-			ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
-				Connector: &common.ConnectorConfig{
-					Driver: "memory",
-					Memory: &common.MemoryConnectorConfig{
-						MaxItems:     100_000,
-						MaxTotalSize: "1MB",
-					},
-				},
-			})
-			require.NoError(t, err)
-
-			upsReg := upstream.NewUpstreamsRegistry(
-				ctx, &log.Logger, "prjA", tt.upstreams,
-				ssr, nil, vr, pr, nil, mt, 1*time.Second,
-			)
-
-			// Create consensus config with new preference flags
-			consensusConfig := &common.ConsensusPolicyConfig{
-				MaxParticipants:         tt.maxParticipants,
-				AgreementThreshold:      tt.agreementThreshold,
-				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+			expectedError: &expectedError{code: common.ErrCodeConsensusLowParticipants, contains: "not enough participants"},
+			setupFn: func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry) {
+				ups := reg.GetNetworkUpstreams(ctx, util.EvmNetworkId(123))
+				ups[0].EvmStatePoller().SuggestLatestBlock(300)
+				ups[1].EvmStatePoller().SuggestLatestBlock(100)
+			},
+		},
+		{
+			name:        "non_empty_over_errors_and_empty_mixed_below_threshold",
+			description: "Below threshold mix: 2 identical errors, 1 empty, 1 non-empty; AcceptMostCommon + preferNonEmpty should select the non-empty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
 				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-			}
-
-			// Set preference flags if specified
-			if tt.preferNonEmpty != nil {
-				consensusConfig.PreferNonEmpty = tt.preferNonEmpty
-			}
-			if tt.preferLargerResponses != nil {
-				consensusConfig.PreferLargerResponses = tt.preferLargerResponses
-			}
-
-			// Create network with consensus policy
-			ntw, err := NewNetwork(
-				ctx, &log.Logger, "prjA",
-				&common.NetworkConfig{
-					Architecture: common.ArchitectureEvm,
-					Evm: &common.EvmNetworkConfig{
-						ChainId: 998,
-					},
-					Failsafe: []*common.FailsafeConfig{
-						{
-							MatchMethod: "*",
-							Consensus:   consensusConfig,
-						},
-					},
-				},
-				nil, upsReg, mt,
-			)
-			require.NoError(t, err)
-
-			err = upsReg.Bootstrap(ctx)
-			require.NoError(t, err)
-			err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(998))
-			require.NoError(t, err)
-
-			// Setup mock responses
-			for i, upstream := range tt.upstreams {
-				gock.New(upstream.Endpoint).
-					Post("/").
-					Times(1). // Each upstream should only be called once
-					Reply(200).
-					SetHeader("Content-Type", "application/json").
-					JSON(tt.mockResponses[i])
-			}
-
-			// Make request
-			reqBytes, err := json.Marshal(tt.request)
-			require.NoError(t, err)
-
-			fakeReq := common.NewNormalizedRequest(reqBytes)
-			resp, err := ntw.Forward(ctx, fakeReq)
-
-			// Verify results
-			require.NoError(t, err)
-			require.NotNil(t, resp)
-
-			// Get the JSON-RPC response to analyze
-			jr, err := resp.JsonRpcResponse(ctx)
-			require.NoError(t, err)
-
-			// Convert response to JSON for analysis
-			respBytes := jr.Result
-			respStr := string(respBytes)
-
-			// Verify response size/content based on expectations
-			if tt.expectedResponseSize == "larger" {
-				// Should contain the complete response with logs and transfer event
-				assert.Contains(t, respStr, "logs", "Expected larger response to contain logs")
-				assert.Contains(t, respStr, "logsBloom", "Expected larger response to contain logsBloom")
-				if tt.expectedResponseContains != "" {
-					assert.Contains(t, respStr, tt.expectedResponseContains,
-						"Expected larger response to contain %s", tt.expectedResponseContains)
-				}
-			} else if tt.expectedResponseSize == "smaller" {
-				// Should be the incomplete response without logs
-				assert.NotContains(t, respStr, "logs", "Expected smaller response to not contain logs")
-				assert.NotContains(t, respStr, "logsBloom", "Expected smaller response to not contain logsBloom")
-			}
-
-			// Log the actual response size for debugging
-			responseSize, _ := jr.Size(ctx)
-			t.Logf("Response size: %d bytes, Expected: %s", responseSize, tt.expectedResponseSize)
-			t.Logf("Response preview: %.200s...", respStr)
-		})
-	}
-}
-
-func TestNetwork_ConsensusPolicy_ErrorsShouldNotWinWithSizePreference(t *testing.T) {
-	// This test reproduces the exact issue from production logs where:
-	// - 2 upstreams returned empty results ([] with resultSize: 2)
-	// - 1 upstream returned an error (ErrUpstreamsExhausted)
-	// - The ERROR was incorrectly winning due to size preference logic
-	// - Empty responses should win by COUNT (2 > 1), not errors winning by size
-
-	util.ResetGock()
-	defer util.ResetGock()
-	util.SetupMocksForEvmStatePoller()
-	defer util.AssertNoPendingMocks(t, 0)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	// Setup network infrastructure (following the same pattern as existing tests)
-	mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
-	vr := thirdparty.NewVendorsRegistry()
-	pr, err := thirdparty.NewProvidersRegistry(&log.Logger, vr, []*common.ProviderConfig{}, nil)
-	require.NoError(t, err)
-
-	ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
-		Connector: &common.ConnectorConfig{
-			Driver: "memory",
-			Memory: &common.MemoryConnectorConfig{
-				MaxItems:     100_000,
-				MaxTotalSize: "1MB",
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
 			},
-		},
-	})
-	require.NoError(t, err)
-
-	// Create upstreams
-	upstreams := []*common.UpstreamConfig{
-		{
-			Id:       "upstream1-empty",
-			Type:     common.UpstreamTypeEvm,
-			Endpoint: "http://upstream1.localhost",
-			Evm:      &common.EvmUpstreamConfig{ChainId: 42161},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess("0xnonempty")},
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xnonempty\""},
 		},
 		{
-			Id:       "upstream2-empty",
-			Type:     common.UpstreamTypeEvm,
-			Endpoint: "http://upstream2.localhost",
-			Evm:      &common.EvmUpstreamConfig{ChainId: 42161},
-		},
-		{
-			Id:       "upstream3-error",
-			Type:     common.UpstreamTypeEvm,
-			Endpoint: "http://upstream3.localhost",
-			Evm:      &common.EvmUpstreamConfig{ChainId: 42161},
-		},
-	}
-
-	upsReg := upstream.NewUpstreamsRegistry(
-		ctx, &log.Logger, "prjA", upstreams,
-		ssr, nil, vr, pr, nil, mt, 1*time.Second,
-	)
-
-	// Create consensus config matching EXACT production settings from the user
-	consensusConfig := &common.ConsensusPolicyConfig{
-		MaxParticipants:         3,
-		AgreementThreshold:      2,
-		DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-		LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError, // Production uses returnError, not acceptMostCommon!
-		PreferLargerResponses:   pointer(true),                                      // Explicitly enable this to trigger the bug
-		PreferNonEmpty:          pointer(true),                                      // Explicitly enable this too
-	}
-
-	// Mock responses: 2 upstreams return empty array, 1 fails
-	emptyResponse := map[string]interface{}{
-		"jsonrpc": "2.0",
-		"id":      1,
-		"result":  []interface{}{}, // Empty array - exactly like the production log
-	}
-
-	// Mock upstream1 and upstream2 to return empty results
-	gock.New("http://upstream1.localhost").
-		Post("/").
-		Times(1).
-		Reply(200).
-		JSON(emptyResponse)
-
-	gock.New("http://upstream2.localhost").
-		Post("/").
-		Times(1).
-		Reply(200).
-		JSON(emptyResponse)
-
-	// Mock upstream3 to return an error (simulating ErrUpstreamsExhausted scenario)
-	// This time, let's make it return a larger "size" error by making it a longer message
-	gock.New("http://upstream3.localhost").
-		Post("/").
-		Times(1).
-		Reply(500).
-		JSON(map[string]interface{}{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"error": map[string]interface{}{
-				"code":    -32000,
-				"message": "upstream exhausted: all upstream attempts failed with very long error message that should make this response larger than the empty array responses if size preference is incorrectly applied to errors which would be a bug that needs to be fixed",
-			},
-		})
-
-	// Create network with consensus policy
-	ntw, err := NewNetwork(
-		ctx, &log.Logger, "prjA",
-		&common.NetworkConfig{
-			Architecture: common.ArchitectureEvm,
-			Evm: &common.EvmNetworkConfig{
-				ChainId: 42161, // Arbitrum
-			},
-			Failsafe: []*common.FailsafeConfig{
-				{
-					MatchMethod: "*",
-					Consensus:   consensusConfig,
-				},
-			},
-		},
-		nil, upsReg, mt,
-	)
-	require.NoError(t, err)
-
-	err = upsReg.Bootstrap(ctx)
-	require.NoError(t, err)
-	err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(42161))
-	require.NoError(t, err)
-
-	// Make the request (using eth_getLogs like in the production issue)
-	request := map[string]interface{}{
-		"method": "eth_getLogs",
-		"params": []interface{}{
-			map[string]interface{}{
-				"address": []interface{}{
-					"0x3bf6d2790779bd1d9526c425be9f92a0900ea64c",
-				},
-				"fromBlock": "0x15c4bf51",
-				"toBlock":   "0x15c4bf52",
-				"topics": []interface{}{
-					"0x2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d",
-				},
-			},
-		},
-	}
-
-	reqBytes, err := json.Marshal(request)
-	require.NoError(t, err)
-
-	fakeReq := common.NewNormalizedRequest(reqBytes)
-	resp, err := ntw.Forward(ctx, fakeReq)
-
-	// If the bug exists, this should fail because the error would win due to larger size
-	// causing a consensus dispute, since we have disputeBehavior: returnError
-
-	if err != nil {
-		// If we get an error, it might be because the error response incorrectly won
-		// due to size preference, causing a dispute
-		t.Logf("❌ Got error (this might indicate the size preference bug): %v", err)
-
-		// Check if it's a consensus dispute error
-		if strings.Contains(err.Error(), "consensus") || strings.Contains(err.Error(), "dispute") {
-			t.Logf("🐛 CONFIRMED: Size preference bug exists - error won over empty responses, causing dispute")
-			t.Logf("This proves that size preference is incorrectly being applied to error responses")
-			// This is the bug we want to fix
-		}
-
-		// For now, let's expect this to fail to demonstrate the bug
-		require.NoError(t, err, "BUG REPRODUCTION: Error won due to size preference - this is the bug we need to fix")
-	}
-
-	require.NotNil(t, resp)
-
-	// Get the JSON-RPC response to analyze
-	jr, err := resp.JsonRpcResponse(ctx)
-	require.NoError(t, err)
-
-	// Verify we got the empty array result (not an error)
-	respBytes := jr.Result
-	respStr := string(respBytes)
-
-	// The key assertion: we should get empty array [], NOT an error
-	// This proves that empty responses won by count (2 > 1), not error winning by size
-	assert.Equal(t, "[]", respStr, "Should get empty array result, proving empty responses won by count over error")
-
-	// Additional verification: ensure it's actually an empty array
-	var resultArray []interface{}
-	err = json.Unmarshal(respBytes, &resultArray)
-	require.NoError(t, err, "Result should be a valid empty array")
-	assert.Empty(t, resultArray, "Result array should be empty")
-
-	// Log for debugging
-	responseSize, _ := jr.Size(ctx)
-	t.Logf("Final response size: %d bytes", responseSize)
-	t.Logf("Final response: %s", respStr)
-	t.Logf("✅ Empty responses correctly won by count (2 > 1) over error, size preference did not apply to errors")
-}
-
-func TestConsensusErrorGroupsOverrideEmptyResponsesBug(t *testing.T) {
-	// This test demonstrates the bug where error groups always win over empty responses
-	// when preferNonEmpty is true, due to an early return in determineWinner() method.
-	// Empty responses with more votes should beat error responses with fewer votes.
-
-	util.ResetGock()
-	defer util.ResetGock()
-	util.SetupMocksForEvmStatePoller()
-	defer util.AssertNoPendingMocks(t, 0)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer func() {
-		cancel()
-		time.Sleep(100 * time.Millisecond)
-	}()
-
-	// Setup network infrastructure
-	mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
-	vr := thirdparty.NewVendorsRegistry()
-	pr, err := thirdparty.NewProvidersRegistry(&log.Logger, vr, []*common.ProviderConfig{}, nil)
-	require.NoError(t, err)
-
-	ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
-		Connector: &common.ConnectorConfig{
-			Driver: "memory",
-			Memory: &common.MemoryConnectorConfig{
-				MaxItems:     100_000,
-				MaxTotalSize: "1MB",
-			},
-		},
-	})
-	require.NoError(t, err)
-
-	// Create upstreams
-	upstreams := []*common.UpstreamConfig{
-		{
-			Id:       "empty1",
-			Type:     common.UpstreamTypeEvm,
-			Endpoint: "http://empty1.localhost",
-			Evm:      &common.EvmUpstreamConfig{ChainId: 1},
-		},
-		{
-			Id:       "empty2",
-			Type:     common.UpstreamTypeEvm,
-			Endpoint: "http://empty2.localhost",
-			Evm:      &common.EvmUpstreamConfig{ChainId: 1},
-		},
-		{
-			Id:       "error1",
-			Type:     common.UpstreamTypeEvm,
-			Endpoint: "http://error1.localhost",
-			Evm:      &common.EvmUpstreamConfig{ChainId: 1},
-		},
-	}
-
-	upsReg := upstream.NewUpstreamsRegistry(
-		ctx, &log.Logger, "prjA", upstreams,
-		ssr, nil, vr, pr, nil, mt, 1*time.Second,
-	)
-
-	// Create consensus config with preferNonEmpty = true
-	// This triggers the bug where error groups override empty responses
-	consensusConfig := &common.ConsensusPolicyConfig{
-		MaxParticipants:         3,
-		AgreementThreshold:      2,
-		DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
-		LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-		PreferNonEmpty:          pointer(true),  // This is key to trigger the bug
-		PreferLargerResponses:   pointer(false), // Disable size preference to focus on the bug
-	}
-
-	// Setup mock responses: 2 empty responses (should win) vs 1 error response (should lose)
-	emptyResponse := map[string]interface{}{
-		"jsonrpc": "2.0",
-		"id":      1,
-		"result":  []interface{}{}, // Empty result
-	}
-
-	// Mock 2 upstreams to return empty results (should win by count)
-	gock.New("http://empty1.localhost").
-		Post("/").
-		Times(1).
-		Reply(200).
-		JSON(emptyResponse)
-
-	gock.New("http://empty2.localhost").
-		Post("/").
-		Times(1).
-		Reply(200).
-		JSON(emptyResponse)
-
-	// Mock 1 upstream to return an error (should lose by count)
-	gock.New("http://error1.localhost").
-		Post("/").
-		Times(1).
-		Reply(500).
-		JSON(map[string]interface{}{
-			"jsonrpc": "2.0",
-			"id":      1,
-			"error": map[string]interface{}{
-				"code":    -32000,
-				"message": "execution reverted",
-			},
-		})
-
-	// Create network
-	ntw, err := NewNetwork(
-		ctx, &log.Logger, "prjA",
-		&common.NetworkConfig{
-			Architecture: common.ArchitectureEvm,
-			Evm: &common.EvmNetworkConfig{
-				ChainId: 1,
-			},
-			Failsafe: []*common.FailsafeConfig{
-				{
-					MatchMethod: "*",
-					Consensus:   consensusConfig,
-				},
-			},
-		},
-		nil, upsReg, mt,
-	)
-	require.NoError(t, err)
-
-	err = upsReg.Bootstrap(ctx)
-	require.NoError(t, err)
-	err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(1))
-	require.NoError(t, err)
-
-	// Make request
-	request := map[string]interface{}{
-		"method": "eth_getLogs",
-		"params": []interface{}{
-			map[string]interface{}{
-				"fromBlock": "0x1",
-				"toBlock":   "0x2",
-			},
-		},
-	}
-
-	reqBytes, err := json.Marshal(request)
-	require.NoError(t, err)
-
-	fakeReq := common.NewNormalizedRequest(reqBytes)
-	resp, err := ntw.Forward(ctx, fakeReq)
-
-	// BUG EXPECTATION: Due to the early return bug in determineWinner(),
-	// error responses will win over empty responses regardless of vote count
-	// when preferNonEmpty is true.
-
-	if err != nil {
-		// If we get an error, the bug might be present
-		t.Logf("BUG DETECTED: Got error instead of empty response: %v", err)
-
-		// Check if it's related to consensus
-		if strings.Contains(err.Error(), "consensus") || strings.Contains(err.Error(), "dispute") {
-			t.Logf("🐛 BUG CONFIRMED: Error response incorrectly won over empty responses with more votes")
-			t.Logf("This happens because determineWinner() has early return after error groups, skipping empty groups")
-
-			// This test should fail to demonstrate the bug exists
-			t.Fatalf("BUG REPRODUCTION: Error responses (1 vote) incorrectly beat empty responses (2 votes) due to early return bug")
-		}
-
-		require.NoError(t, err, "Unexpected error during consensus")
-	}
-
-	require.NotNil(t, resp)
-
-	// Get the JSON-RPC response
-	jr, err := resp.JsonRpcResponse(ctx)
-	require.NoError(t, err)
-
-	// Verify we got the empty array result (correct behavior)
-	respBytes := jr.Result
-	respStr := string(respBytes)
-
-	// CORRECT BEHAVIOR: Empty responses should win by count (2 > 1)
-	assert.Equal(t, "[]", respStr, "Empty responses should win by count over error responses")
-
-	// Additional verification
-	var resultArray []interface{}
-	err = json.Unmarshal(respBytes, &resultArray)
-	require.NoError(t, err, "Result should be a valid empty array")
-	assert.Empty(t, resultArray, "Result array should be empty")
-
-	// Log for debugging
-	responseSize, _ := jr.Size(ctx)
-	t.Logf("Final response size: %d bytes", responseSize)
-	t.Logf("Final response: %s", respStr)
-	t.Logf("✅ Empty responses correctly won by count (2 > 1) over error response (1)")
-}
-
-func TestConsensusEmptyVsErrorsPriority(t *testing.T) {
-	// This test verifies the priority ordering where empty successful responses
-	// are always preferred over errors, regardless of vote counts.
-	// Priority: non-empty > empty > consensus-valid errors > generic errors
-	// Scenario: 3 errors vs 1 empty response -> empty response wins by priority
-
-	util.ConfigureTestLogger()
-	util.ResetGock()
-	defer util.ResetGock()
-	util.SetupMocksForEvmStatePoller()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	// Test both AcceptMostCommonValidResult and AcceptMostCommonValidResult behaviors
-	testCases := []struct {
-		name            string
-		disputeBehavior common.ConsensusDisputeBehavior
-	}{
-		{
-			name:            "AcceptMostCommonValidResult",
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-		},
-		{
-			name:            "AcceptMostCommonValidResult",
-			disputeBehavior: common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Reset gock for each test case
-			util.ResetGock()
-			defer util.ResetGock()
-			util.SetupMocksForEvmStatePoller()
-
-			// Create 4 upstreams
-			upstreams := []*common.UpstreamConfig{
-				{Id: "error1", Type: common.UpstreamTypeEvm, Endpoint: "http://error1.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
-				{Id: "error2", Type: common.UpstreamTypeEvm, Endpoint: "http://error2.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
-				{Id: "error3", Type: common.UpstreamTypeEvm, Endpoint: "http://error3.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
-				{Id: "empty1", Type: common.UpstreamTypeEvm, Endpoint: "http://empty1.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
-			}
-
-			// Create error response for missing data
-			errorResponse := map[string]interface{}{
-				"jsonrpc": "2.0",
-				"id":      1,
-				"error": map[string]interface{}{
-					"code":    -32000,
-					"message": "missing historical data",
-				},
-			}
-
-			// Create empty response
-			emptyResponse := map[string]interface{}{
-				"jsonrpc": "2.0",
-				"id":      1,
-				"result":  []interface{}{},
-			}
-
-			// Mock 3 upstreams returning missing data error
-			gock.New("http://error1.localhost").
-				Post("").
-				Persist().
-				Reply(200).
-				JSON(errorResponse)
-
-			gock.New("http://error2.localhost").
-				Post("").
-				Persist().
-				Reply(200).
-				JSON(errorResponse)
-
-			gock.New("http://error3.localhost").
-				Post("").
-				Persist().
-				Reply(200).
-				JSON(errorResponse)
-
-			// Mock 1 upstream returning empty array
-			gock.New("http://empty1.localhost").
-				Post("").
-				Persist().
-				Reply(200).
-				JSON(emptyResponse)
-
-			// Create network with consensus policy
-			preferNonEmpty := true
-			consensusConfig := &common.ConsensusPolicyConfig{
+			name:        "consensus_error_over_empty_at_threshold_no_non_empty",
+			description: "3 consensus-valid errors vs 1 empty; AcceptMostCommon (+ PreferNonEmpty) should return the error since no non-empty exists",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
 				MaxParticipants:         4,
 				AgreementThreshold:      2,
-				PreferNonEmpty:          &preferNonEmpty, // This should prefer empty over errors
-				DisputeBehavior:         tc.disputeBehavior,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
 				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-			}
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(3, "execution reverted")},
+				{status: 200, body: jsonRpcError(3, "execution reverted")},
+				{status: 200, body: jsonRpcError(3, "execution reverted")},
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+			},
+			expectedCalls: []int{1, 1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "execution reverted"},
+		},
+		{
+			name:        "error_has_higher_count_but_non_empty_preferred_above_threshold",
+			description: "Above threshold both non-empty and error groups; error has higher count; AcceptMostCommon + preferNonEmpty should still choose non-empty",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xaaa\""},
+		},
+		{
+			name:        "participant_requirement_more_than_available_low_participants",
+			description: "Agreement threshold exceeds available valid participants -> low participants error",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xdata")},
+				{status: 200, body: jsonRpcSuccess("0xdata")},
+			},
+			expectedCalls: []int{1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusLowParticipants, contains: "not enough participants"},
+		},
+		{
+			name:        "minority_success_over_majority_execution_errors_with_preference",
+			description: "Prefer non-empty success when below threshold vs multiple execution errors under AcceptMostCommon+preferNonEmpty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: out of gas")},
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: out of gas")},
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: out of gas")},
+				{status: 200, body: jsonRpcSuccess("0xsuccess")},
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xsuccess\""},
+		},
+		{
+			name:        "return_error_behavior_different_execution_errors_returns_one_error",
+			description: "With ReturnError behavior and different execution errors, return one of the execution errors",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: reason A")},
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: reason B")},
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: reason C")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "execution reverted"},
+		},
+		{
+			name:        "mixed_infra_and_execution_errors_consensus_on_execution_error",
+			description: "1 infra error (-32603) and 2 identical execution errors (-32000) -> consensus on execution error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32603, "network timeout")},
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: insufficient balance")},
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: insufficient balance")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "execution reverted"},
+		},
+		{
+			name:        "agreed_invalid_params_error",
+			description: "All participants return the same JSON-RPC invalid params error -> return that error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "Invalid params"},
+		},
+		{
+			name:        "agreed_method_not_found_error",
+			description: "All participants return method not found error -> return that error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32601, "Method not found")},
+				{status: 200, body: jsonRpcError(-32601, "Method not found")},
+				{status: 200, body: jsonRpcError(-32601, "Method not found")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "Method not found"},
+		},
+		{
+			name:        "agreed_missing_data_error",
+			description: "All participants return missing data (-32014) -> return that error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32014, "requested data is not available")},
+				{status: 200, body: jsonRpcError(-32014, "requested data is not available")},
+				{status: 200, body: jsonRpcError(-32014, "requested data is not available")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "requested data is not available"},
+		},
+		{
+			name:        "agreed_execution_reverted_error",
+			description: "All participants return execution reverted (code 3) -> return that error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(3, "execution reverted")},
+				{status: 200, body: jsonRpcError(3, "execution reverted")},
+				{status: 200, body: jsonRpcError(3, "execution reverted")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "execution reverted"},
+		},
+		{
+			name:        "mixed_errors_returns_dispute",
+			description: "Different JSON-RPC errors from participants below threshold -> dispute",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+				{status: 200, body: jsonRpcError(-32601, "Method not found")},
+				{status: 200, body: jsonRpcError(-32000, "Internal error")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusDispute, contains: "not enough agreement"},
+		},
+		{
+			name:        "majority_error_over_single_success_returns_error",
+			description: "Two identical JSON-RPC errors vs one success at threshold -> return the agreed error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+				{status: 200, body: jsonRpcSuccess("0x1234"), delay: 300 * time.Millisecond},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "Invalid params"},
+		},
+		{
+			name:        "errors_should_not_win_by_size_preference_return_error",
+			description: "PreferLargerResponses enabled with ReturnError: 2 empty vs 1 large error -> empty wins, errors do not win by size",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.TRUE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcError(-32000, strings.Repeat("E", 500))},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "[]"},
+		},
+		{
+			name:        "prefer_non_empty_over_consensus_error_at_threshold_accept_most_common",
+			description: "PreferNonEmpty + AcceptMostCommon: consensus error meets threshold but single non-empty exists -> select non-empty",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+				{status: 200, body: jsonRpcError(-32602, "Invalid params")},
+				{status: 200, body: jsonRpcSuccess("0x1234")},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0x1234"`},
+		},
+		{
+			name:        "accept_most_common_below_threshold_empty_unique_no_preference",
+			description: "Below threshold unique empty leader vs one non-empty and infra error; no non-empty preference -> select empty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},              // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},              // empty 2 (unique leader below threshold)
+				{status: 200, body: jsonRpcSuccess("0xnonempty")},                 // non-empty 1
+				{status: 200, body: jsonRpcError(-32603, "server error (infra)")}, // infra error
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "[]"},
+		},
+		{
+			name:        "empty_meets_threshold_short_circuits_without_preference",
+			description: "Empty meets threshold quickly; no non-empty preference -> short-circuit to empty",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                         // empty fast 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                         // empty fast 2 (meets threshold)
+				{status: 200, body: jsonRpcSuccess("0xlate"), delay: 300 * time.Millisecond}, // slow third
+			},
+			expectedCalls: []int{1, 1, 0},
+			expectedResult: &expectedResult{
+				jsonRpcResult: "[]",
+				check: func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration) {
+					assert.LessOrEqual(t, duration, 100*time.Millisecond, "should short-circuit on empty when no non-empty preference")
+				},
+			},
+		},
+		{
+			name:        "accept_most_common_below_threshold_empty_unique_with_preference_non_empty",
+			description: "Below threshold unique empty leader vs single non-empty; with preferNonEmpty select non-empty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},              // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},              // empty 2 (unique leader below threshold)
+				{status: 200, body: jsonRpcSuccess("0xnonempty")},                 // non-empty 1
+				{status: 200, body: jsonRpcError(-32603, "server error (infra)")}, // infra error
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xnonempty\""},
+		},
+		{
+			name:        "threshold_2_two_identical_success",
+			description: "2 participants, threshold 2, both non-empty identical -> success",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xabc")},
+				{status: 200, body: jsonRpcSuccess("0xabc")},
+			},
+			expectedCalls:  []int{1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xabc\""},
+		},
+		{
+			name:        "threshold_2_two_different_dispute",
+			description: "2 participants, threshold 2, different non-empty -> dispute",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+				{status: 200, body: jsonRpcSuccess("0xbbb")},
+			},
+			expectedCalls: []int{1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusDispute, contains: "not enough agreement"},
+		},
+		{
+			name:        "threshold_2_two_empty_success",
+			description: "2 participants, threshold 2, both empty -> success with empty",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+			},
+			expectedCalls:  []int{1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "[]"},
+		},
+		{
+			name:        "above_threshold_empty_wins_without_preference",
+			description: "Empty group meets threshold with higher count; no non-empty preference -> select empty",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 3 (meets threshold)
+				{status: 200, body: jsonRpcSuccess("0xnonempty")},    // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0xnonempty")},    // non-empty 2
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "[]"},
+		},
+		{
+			name:        "low_participants_return_error_when_only_empty_valid",
+			description: "Low participants with only empty valid responses and ReturnError -> low participants error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      4, // force low participants
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusLowParticipants, contains: "not enough participants"},
+		},
+		{
+			name:        "low_participants_accept_most_common_empty_vs_errors",
+			description: "Low participants with AcceptMostCommon: empty valid beats infra errors",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      4, // force low participants
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                // empty valid
+				{status: 200, body: jsonRpcError(-32603, "upstream error (infra)")}, // infra error
+				{status: 200, body: jsonRpcError(-32603, "upstream error (infra)")}, // infra error
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "[]"},
+		},
+		{
+			name:        "non_consensus_error_vs_non_empty_below_threshold_prefer_non_empty",
+			description: "Below threshold with a non-consensus (infra) error vs a single non-empty: prefer the valid non-empty",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32603, "internal server error")},
+				{status: 200, body: jsonRpcError(-32603, "internal server error")},
+				{status: 200, body: jsonRpcSuccess("0xres")},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xres\""},
+		},
+		{
+			name:        "below_threshold_error_unique_vs_single_non_empty_prefer_non_empty",
+			description: "Below threshold unique consensus-valid error vs single non-empty; prefer non-empty under AcceptMostCommon + preferNonEmpty",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcSuccess("0xabc")},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xabc\""},
+		},
+		{
+			name:        "above_threshold_non_empty_vs_error_counts_choose_non_empty_when_higher",
+			description: "Above threshold both non-empty and error groups; higher count wins. If tie without preference, dispute; with preferNonEmpty=true, choose non-empty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+				{status: 200, body: jsonRpcError(-32000, "call exception")},
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xaaa\""},
+		},
+		{
+			name:        "ignore_fields_enable_consensus_on_block_timestamp",
+			description: "Ignore block timestamp differences to achieve consensus",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:    3,
+				AgreementThreshold: 2,
+				DisputeBehavior:    common.ConsensusDisputeBehaviorReturnError,
+				IgnoreFields: map[string][]string{
+					"eth_getBlockByNumber": {"timestamp"},
+				},
+			},
+			requestMethod: "eth_getBlockByNumber",
+			requestParams: []interface{}{"0x1", false},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"number": "0x1", "hash": "0xabc123", "timestamp": "0xaaa", "gasLimit": "0x1", "gasUsed": "0x1"})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"number": "0x1", "hash": "0xabc123", "timestamp": "0xbbb", "gasLimit": "0x1", "gasUsed": "0x1"})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"number": "0x1", "hash": "0xdef456", "timestamp": "0xccc", "gasLimit": "0x1", "gasUsed": "0x1"})},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{contains: "\"hash\":\"0xabc123\""},
+		},
+		{
+			name:        "ignore_fields_dispute_when_core_fields_differ",
+			description: "Even with timestamp ignored, different core fields should cause dispute",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:    3,
+				AgreementThreshold: 2,
+				DisputeBehavior:    common.ConsensusDisputeBehaviorReturnError,
+				IgnoreFields: map[string][]string{
+					"eth_getBlockByNumber": {"timestamp"},
+				},
+			},
+			requestMethod: "eth_getBlockByNumber",
+			requestParams: []interface{}{"0x1", false},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"number": "0x1", "hash": "0xaaa111", "timestamp": "0x1", "gasLimit": "0x1", "gasUsed": "0x1"})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"number": "0x2", "hash": "0xbbb222", "timestamp": "0x2", "gasLimit": "0x1", "gasUsed": "0x1"})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"number": "0x3", "hash": "0xccc333", "timestamp": "0x3", "gasLimit": "0x1", "gasUsed": "0x1"})},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusDispute, contains: "not enough agreement"},
+		},
+		{
+			name:        "accept_most_common_below_threshold_tie_no_clear_winner_dispute",
+			description: "Below threshold with AcceptMostCommon and tie between different non-empty results -> dispute",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess("0x1")},
+				{status: 200, body: jsonRpcSuccess("0x2")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusDispute, contains: "not enough agreement"},
+		},
+		{
+			name:        "low_participants_accept_most_common_empty_only",
+			description: "Low participants with AcceptMostCommon and only empty valid responses present -> return best empty",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcError(-32603, "server error")},
+				{status: 200, body: jsonRpcError(-32603, "server error")},
+				{status: 200, body: jsonRpcError(-32603, "server error")},
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{contains: "[]"},
+		},
+		{
+			name:        "prefer_transaction_receipt_with_logs_over_empty_logs",
+			description: "Prefer receipt with non-empty logs over receipts with empty logs under AcceptMostCommon",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			requestMethod: "eth_getTransactionReceipt",
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"logs": []interface{}{}})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"logs": []interface{}{}})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"logs": []interface{}{}})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{"logs": []interface{}{map[string]interface{}{"address": "0x1"}}})},
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{contains: "\"address\""},
+		},
+		{
+			name:        "dispute_with_empty_preference_return_error",
+			description: "Mixed empty and two different non-empty below threshold with ReturnError should dispute",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess("0x123")},
+				{status: 200, body: jsonRpcSuccess("0x456")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusDispute, contains: "not enough agreement"},
+		},
+		{
+			name:        "prefer_non_empty_3v1_with_accept_behavior",
+			description: "3 empty vs 1 non-empty; with AcceptMostCommon + preferNonEmpty, choose the non-empty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess("0x789")},
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x789\""},
+		},
+		{
+			name:        "emptyish_consensus_overridden_by_meaningful_data",
+			description: "Two emptyish (null) vs one non-empty; AcceptMostCommon + preferNonEmpty should pick the non-empty",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess("0x123")},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x123\""},
+		},
+		{
+			name:        "all_emptyish_responses_return_emptyish_consensus",
+			description: "All participants return emptyish (null); should return emptyish consensus",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess(nil)},
+				{status: 200, body: jsonRpcSuccess(nil)},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "null"},
+		},
+		{
+			name:        "emptyish_zero_hex_overridden_by_non_empty_no_short_circuit",
+			description: "Two zero-hex emptyish meet threshold first; with preferNonEmpty we should wait and select the non-empty",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x0")},
+				{status: 200, body: jsonRpcSuccess("0x0")},
+				{status: 200, body: jsonRpcSuccess("0x789"), delay: 300 * time.Millisecond},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedResult: &expectedResult{
+				jsonRpcResult: "\"0x789\"",
+				check: func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration) {
+					assert.GreaterOrEqual(t, duration, 300*time.Millisecond, "should not short-circuit when emptyish leads and preferNonEmpty is enabled")
+				},
+			},
+		},
+		{
+			name:        "mixed_emptyish_types_overridden_by_non_empty",
+			description: "Empty array and empty object vs one non-empty; prefer non-empty below threshold",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess(map[string]interface{}{})},
+				{status: 200, body: jsonRpcSuccess("0xdef")},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xdef\""},
+		},
+		{
+			name:        "successful_consensus_2_of_3_prefer_larger_responses_enabled",
+			description: "A simple successful consensus where 2 out of 3 upstreams agree.",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcSuccess("0x7b"), delay: 300 * time.Millisecond},
+			},
+			expectedCalls: []int{1, 1, 1}, // will not short circuit because prefer larger responses is enabled
+			expectedResult: &expectedResult{
+				jsonRpcResult: `"0x7a"`,
+				check: func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration) {
+					assert.GreaterOrEqual(t, duration, 300*time.Millisecond, "response should take more than 300ms and no short circuit")
+				},
+			},
+		},
+		{
+			name:        "successful_consensus_2_of_3_prefer_larger_responses_disabled",
+			description: "A simple successful consensus where 2 out of 3 upstreams agree.",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcSuccess("0x7b"), delay: 300 * time.Millisecond},
+			},
+			expectedCalls: []int{1, 1, 0}, // last request is not sent due to short circuit
+			expectedResult: &expectedResult{
+				jsonRpcResult: `"0x7a"`,
+				check: func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration) {
+					assert.Less(t, duration, 100*time.Millisecond, "response should short circuit not wait for the slowest upstream")
+				},
+			},
+		},
+		{
+			name:        "low_participants_only_one_participant",
+			description: "A single participant returns a result, but no consensus is reached.",
+			upstreams:   createTestUpstreams(1),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls: []int{1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusLowParticipants,
+				contains: "not enough participants",
+			},
+		},
+		{
+			name:        "prefer_non_empty_over_empty_above_threshold_accept_most_common",
+			description: "Prefer non-empty when empty meets threshold (3) but non-empty has 2; AcceptMostCommon allows selecting the non-empty",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x11")},          // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0x11")},          // non-empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 3 -> meets threshold
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0x11"`},
+		},
+		{
+			name:        "prefer_non_empty_but_return_error_when_empty_wins_with_threshold",
+			description: "Prefer non-empty enabled, but with ReturnError behavior and empty meeting threshold, return dispute error",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x22")},          // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0x22")},          // non-empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 3 -> meets threshold
+			},
+			expectedCalls: []int{1, 1, 1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusDispute,
+				contains: "not enough agreement among responses",
+			},
+		},
+		{
+			name:        "prefer_non_empty_when_both_groups_above_threshold",
+			description: "Two groups empty and non-empty both meet threshold; preferNonEmpty selects non-empty",
+			upstreams:   createTestUpstreams(6),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         6,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x33")},          // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0x33")},          // non-empty 2
+				{status: 200, body: jsonRpcSuccess("0x33")},          // non-empty 3 -> meets threshold
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 3 -> meets threshold
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0x33"`},
+		},
+		{
+			name:        "tie_above_threshold_without_preference_results_in_dispute",
+			description: "Two groups empty and non-empty both meet threshold with equal counts; no preference -> dispute",
+			upstreams:   createTestUpstreams(6),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         6,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x44")},          // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0x44")},          // non-empty 2
+				{status: 200, body: jsonRpcSuccess("0x44")},          // non-empty 3 -> meets threshold
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 3 -> meets threshold
+			},
+			expectedCalls: []int{1, 1, 1, 1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusDispute,
+				contains: "not enough agreement among responses",
+			},
+		},
+		{
+			name:        "dispute_on_all_different_responses",
+			description: "A dispute occurs when all upstreams return different results.",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcSuccess("0x8b")},
+				{status: 200, body: jsonRpcSuccess("0x9c")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusDispute,
+				contains: "not enough agreement among responses",
+			},
+		},
+		{
+			name:        "retry_on_error_and_success_on_next_upstream",
+			description: "Retry enabled; second upstream errors; first and third succeed, reaching 2/3 consensus.",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{
+				MaxAttempts: 2,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "cannot query unfinalized data")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedResult: &expectedResult{
+				jsonRpcResult: `"0x7a"`,
+			},
+		},
+		{
+			name:        "retried_failing_usptream_reach_consensus",
+			description: "retry once on intermittently bad upstream and reach consensus if correct result is returned on second attempt",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			retryPolicy: &common.RetryPolicyConfig{
+				MaxAttempts: 2,
+			},
+			// Let harness mock upstream1 once; we’ll override upstream2 via setupFn
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x1")},
+			},
+			expectedCalls: []int{1, 0},
+			expectedResult: &expectedResult{
+				jsonRpcResult: `"0x1"`,
+			},
+			setupFn: func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry) {
+				// upstream2: first error (-32603), then success "0x1"
+				gock.New("http://rpc2.localhost").
+					Post("/").
+					Times(1).
+					Filter(func(request *http.Request) bool {
+						body := util.SafeReadBody(request)
+						return strings.Contains(string(body), "eth_randomMethod")
+					}).
+					Reply(200).
+					SetHeader("Content-Type", "application/json").
+					JSON(jsonRpcError(-32603, "unknown server error"))
+				gock.New("http://rpc2.localhost").
+					Post("/").
+					Times(1).
+					Filter(func(request *http.Request) bool {
+						body := util.SafeReadBody(request)
+						return strings.Contains(string(body), "eth_randomMethod")
+					}).
+					Reply(200).
+					SetHeader("Content-Type", "application/json").
+					JSON(jsonRpcSuccess("0x1"))
+			},
+			expectedPendingMocks: 0,
+		},
+		{
+			name:        "prefer_larger_disabled_above_threshold_different_counts_choose_higher_count",
+			description: "PreferLargerResponses disabled; both non-empty groups >= threshold with different counts -> higher count wins",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x" + strings.Repeat("A", 1000))}, // larger 1 (>= threshold but fewer votes)
+				{status: 200, body: jsonRpcSuccess("0x" + strings.Repeat("A", 1000))}, // larger 2
+				{status: 200, body: jsonRpcSuccess("0x" + strings.Repeat("b", 10))},   // smaller 1
+				{status: 200, body: jsonRpcSuccess("0x" + strings.Repeat("b", 10))},   // smaller 2
+				{status: 200, body: jsonRpcSuccess("0x" + strings.Repeat("b", 10))},   // smaller 3 (higher count)
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x" + strings.Repeat("b", 10) + "\""},
+		},
+		{
+			name:        "prefer_larger_over_error_tie_above_threshold_accept_most_common",
+			description: "PreferLargerResponses enabled; non-empty and consensus-error both meet threshold with equal counts -> choose largest non-empty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x" + strings.Repeat("C", 1000))},     // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0x" + strings.Repeat("C", 1000))},     // non-empty 2 (>= threshold)
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: reason X")}, // error 1
+				{status: 200, body: jsonRpcError(-32000, "execution reverted: reason X")}, // error 2 (>= threshold)
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x" + strings.Repeat("C", 1000) + "\""},
+		},
+		{
+			name:        "some_participants_return_error_success_below_threshold",
+			description: "3 upstreams; threshold 3; two successes and one error -> low participants",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusLowParticipants,
+				contains: "not enough participants",
+			},
+		},
+		{
+			name:        "some_participants_return_error_success_above_threshold",
+			description: "3 upstreams; threshold 2; two successes and one error -> success",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0x7a"`},
+		},
+		{
+			name:        "some_participants_return_error_and_return_error_on_low_participants",
+			description: "Explicit low participants ReturnError; two successes below threshold -> error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusLowParticipants,
+				contains: "not enough participants",
+			},
+		},
+		{
+			name:        "not_enough_upstreams_low_participants_behavior_accept_most_common_valid_result",
+			description: "2 upstreams only; threshold 3; AcceptMostCommon -> return best valid success",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+			},
+			expectedCalls:  []int{1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: `"0x7a"`},
+		},
+		{
+			name:        "agreed_unknown_error_response_on_upstreams",
+			description: "All participants return the same error; consensus on error is returned",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "random error")},
+				{status: 200, body: jsonRpcError(-32000, "random error")},
+				{status: 200, body: jsonRpcError(-32000, "random error")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeUpstreamRequest,
+				contains: "random error",
+			},
+		},
+		{
+			name:        "dispute_behavior_return_error_on_three_different",
+			description: "3 upstreams, all different; ReturnError dispute",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:    3,
+				AgreementThreshold: 2,
+				DisputeBehavior:    common.ConsensusDisputeBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xaaa")},
+				{status: 200, body: jsonRpcSuccess("0xbbb")},
+				{status: 200, body: jsonRpcSuccess("0xccc")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusDispute, contains: "not enough agreement among responses"},
+		},
+		{
+			name:        "low_participants_accept_most_common_picks_non_empty",
+			description: "5 upstreams, only 2 valid respond (empty + non-empty), low participants AcceptMostCommon → non-empty",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},
+				{status: 200, body: jsonRpcSuccess("0xdata")},
+				{status: 200, body: jsonRpcError(-32603, "server error")},
+				{status: 200, body: jsonRpcError(-32603, "server error")},
+				{status: 200, body: jsonRpcError(-32603, "server error")},
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xdata\""},
+		},
+		{
+			name:        "threshold_1_single_response_succeeds",
+			description: "1 upstream only; threshold 1; should succeed with that response",
+			upstreams:   createTestUpstreams(1),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:    1,
+				AgreementThreshold: 1,
+			},
+			mockResponses:  []mockResponse{{status: 200, body: jsonRpcSuccess("0xsingle")}},
+			expectedCalls:  []int{1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0xsingle\""},
+		},
+		{
+			name:        "some_participants_return_error_success_below_threshold",
+			description: "3 upstreams; threshold 3; two successes and one error -> low participants",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusLowParticipants, contains: "not enough participants"},
+		},
+		{
+			name:        "some_participants_return_error_success_above_threshold",
+			description: "3 upstreams; threshold 2; two successes and one error -> success",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls:  []int{1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x7a\""},
+		},
+		{
+			name:        "some_participants_return_error_and_return_error_on_low_participants",
+			description: "Explicit low participants ReturnError; two successes below threshold -> error",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusLowParticipants, contains: "not enough participants"},
+		},
+		{
+			name:        "not_enough_upstreams_low_participants_behavior_accept_most_common_valid_result",
+			description: "2 upstreams only; threshold 3; AcceptMostCommon -> return best valid success",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.FALSE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x7a")},
+				{status: 200, body: jsonRpcError(-32000, "internal error")},
+			},
+			expectedCalls:  []int{1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x7a\""},
+		},
+		{
+			name:        "agreed_unknown_error_response_on_upstreams",
+			description: "All participants return the same error; consensus on error is returned",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcError(-32000, "random error")},
+				{status: 200, body: jsonRpcError(-32000, "random error")},
+				{status: 200, body: jsonRpcError(-32000, "random error")},
+			},
+			expectedCalls: []int{1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeUpstreamRequest, contains: "random error"},
+		},
+		{
+			name:        "accept_most_common_below_threshold_5_participants",
+			description: "5 upstreams, threshold 3; only 2 agree, AcceptMostCommon should return the best group",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xaaa")},                          // up1
+				{status: 200, body: jsonRpcSuccess("0xaaa")},                          // up2 -> best group size 2
+				{status: 200, body: jsonRpcSuccess("0xbbb")},                          // up3
+				{status: 200, body: jsonRpcError(-32000, "cannot query unfinalized")}, // up4 (error group)
+				{status: 200, body: jsonRpcError(-32003, "tx rejected")},              // up5 (another error)
+			},
+			expectedCalls: []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{
+				jsonRpcResult: `"0xaaa"`,
+			},
+		},
+		{
+			name:        "prefer_larger_1_big_below_threshold_vs_3_small_below_threshold_accept_most_common",
+			description: "Prefer larger enabled; below threshold; AcceptMostCommon -> pick 1 largest non-empty; no short circuit",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large non-empty (unique)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("b", 10))},   // small non-empty 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("c", 10))},   // small non-empty 2
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("d", 10))},   // small non-empty 3
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"" + strings.Repeat("A", 2000) + "\""},
+		},
+		{
+			name:        "no_larger_response_preference_threshold_met_short_circuit",
+			description: "non-empty response meeting threshold + PreferLargerResponses=false should short-circuit",
+			upstreams:   createTestUpstreams(3),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         3,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xabc")},
+				{status: 200, body: jsonRpcSuccess("0xabc")},
+				{status: 200, body: jsonRpcSuccess("0xabc"), delay: 300 * time.Millisecond},
+			},
+			expectedCalls: []int{1, 1, 0},
+			expectedResult: &expectedResult{
+				jsonRpcResult: "\"0xabc\"",
+				check: func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration) {
+					assert.LessOrEqual(t, duration, 100*time.Millisecond, "should short-circuit when non-empty response meets threshold and PreferLargerResponses is disabled")
+				},
+			},
+		},
+		{
+			name:        "prefer_non_empty_above_threshold_empty_has_higher_count_accept_most_common",
+			description: "Both groups above threshold, empty has higher count; AcceptMostCommon + preferNonEmpty should pick non-empty",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x55")},          // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0x55")},          // non-empty 2 (>= threshold)
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 2
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 3 (empty has higher count)
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x55\""},
+		},
+		{
+			name:        "prefer_non_empty_above_threshold_empty_arrives_first_no_short_circuit",
+			description: "Empty responses arrive earlier and meet threshold first; with preferNonEmpty we should not short-circuit and wait for non-empty",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                       // empty 1 (fast)
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                       // empty 2 (meets threshold fast)
+				{status: 200, body: jsonRpcSuccess("0x55"), delay: 300 * time.Millisecond}, // non-empty 1 (slow)
+				{status: 200, body: jsonRpcSuccess("0x55"), delay: 300 * time.Millisecond}, // non-empty 2 (slow, meets threshold)
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                       // empty 3
+			},
+			expectedCalls: []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{
+				jsonRpcResult: "\"0x55\"",
+				check: func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration) {
+					assert.GreaterOrEqual(t, duration, 300*time.Millisecond, "should not short-circuit when preferNonEmpty is enabled and empty leads with participants remaining")
+				},
+			},
+		},
+		{
+			name:        "prefer_larger_above_threshold_small_has_higher_count_accept_most_common",
+			description: "Prefer larger enabled; both groups above threshold but smaller has higher count; choose largest",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},   // small 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},   // small 2
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large 2 (>= threshold)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},   // small 3 (small has higher count)
+			},
+			expectedCalls:  []int{1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"" + strings.Repeat("A", 2000) + "\""},
+		},
+		{
+			name:        "prefer_larger_above_threshold_small_has_higher_count_no_short_circuit_when_empty_leads",
+			description: "Empty responses lead early; with PreferNonEmpty enabled we should not short-circuit and wait to select the larger non-empty under AcceptMostCommon",
+			upstreams:   createTestUpstreams(7),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         7,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.TRUE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                                          // empty 1 (fast)
+				{status: 200, body: jsonRpcSuccess([]interface{}{})},                                          // empty 2 (empty meets threshold fast)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},                                  // small 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},                                  // small 2 (small meets threshold)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},                                  // small 3 (small has higher count)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000)), delay: 300 * time.Millisecond}, // large 1 (slow)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000)), delay: 300 * time.Millisecond}, // large 2 (slow, meets threshold)
+			},
+			expectedCalls: []int{1, 1, 1, 1, 1, 1, 1},
+			expectedResult: &expectedResult{
+				jsonRpcResult: "\"" + strings.Repeat("A", 2000) + "\"",
+				check: func(t *testing.T, resp *common.NormalizedResponse, duration time.Duration) {
+					assert.GreaterOrEqual(t, duration, 300*time.Millisecond, "should not short-circuit when preferences are enabled and empty is leading with remaining participants")
+				},
+			},
+		},
+		{
+			name:        "accept_most_common_below_threshold_prefer_non_empty_over_empty",
+			description: "Below threshold; empty uniquely leads; preferNonEmpty selects best non-empty",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferNonEmpty:          &common.TRUE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 1
+				{status: 200, body: jsonRpcSuccess([]interface{}{})}, // empty 2 (unique leader below threshold)
+				{status: 200, body: jsonRpcSuccess("0x1")},           // non-empty 1
+				{status: 200, body: jsonRpcSuccess("0x1")},           // non-empty 2
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"0x1\""},
+		},
+		{
+			name:        "dispute_below_threshold_return_error_5_participants",
+			description: "5 upstreams, threshold 3; only 2 agree, ReturnError should yield dispute",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xaaa")},                          // up1
+				{status: 200, body: jsonRpcSuccess("0xaaa")},                          // up2 -> best group size 2
+				{status: 200, body: jsonRpcSuccess("0xbbb")},                          // up3
+				{status: 200, body: jsonRpcError(-32000, "cannot query unfinalized")}, // up4 (error group)
+				{status: 200, body: jsonRpcError(-32003, "tx rejected")},              // up5 (another error)
+			},
+			expectedCalls: []int{1, 1, 1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusDispute,
+				contains: "not enough agreement among responses",
+			},
+		},
+		{
+			name:        "prefer_larger_1_big_below_threshold_vs_3_small_above_threshold_return_error",
+			description: "Prefer larger enabled; smaller group meets threshold; ReturnError -> dispute (do not accept smaller)",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large non-empty (unique)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},   // small non-empty 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},   // small non-empty 2
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("x", 10))},   // small non-empty 3 (meets threshold)
+			},
+			expectedCalls: []int{1, 1, 1, 1},
+			expectedError: &expectedError{code: common.ErrCodeConsensusDispute, contains: "not enough agreement"},
+		},
+		{
+			name:        "prefer_larger_2_large_meet_threshold_vs_2_small_meet_threshold_accept_most_common",
+			description: "Prefer larger enabled; both groups meet threshold; AcceptMostCommon -> choose larger",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large 2 (meets threshold)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("b", 10))},   // small 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("b", 10))},   // small 2 (meets threshold)
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"" + strings.Repeat("A", 2000) + "\""},
+		},
+		{
+			name:        "prefer_larger_2_large_meet_threshold_vs_2_small_meet_threshold_return_error",
+			description: "Prefer larger enabled; both groups meet threshold; ReturnError -> still choose larger (preference applies)",
+			upstreams:   createTestUpstreams(4),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         4,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferLargerResponses:   &common.TRUE,
+				PreferNonEmpty:          &common.FALSE,
+			},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("A", 2000))}, // large 2 (meets threshold)
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("b", 10))},   // small 1
+				{status: 200, body: jsonRpcSuccess(strings.Repeat("b", 10))},   // small 2 (meets threshold)
+			},
+			expectedCalls:  []int{1, 1, 1, 1},
+			expectedResult: &expectedResult{jsonRpcResult: "\"" + strings.Repeat("A", 2000) + "\""},
+		},
+		{
+			name:        "dispute_below_threshold_return_error_5_participants",
+			description: "5 upstreams, threshold 3; only 2 agree, ReturnError should yield dispute",
+			upstreams:   createTestUpstreams(5),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         5,
+				AgreementThreshold:      3,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+				PreferNonEmpty:          &common.FALSE,
+				PreferLargerResponses:   &common.FALSE,
+			},
+			retryPolicy: &common.RetryPolicyConfig{MaxAttempts: 1},
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0xaaa")},                          // up1
+				{status: 200, body: jsonRpcSuccess("0xaaa")},                          // up2 -> best group size 2
+				{status: 200, body: jsonRpcSuccess("0xbbb")},                          // up3
+				{status: 200, body: jsonRpcError(-32000, "cannot query unfinalized")}, // up4
+				{status: 200, body: jsonRpcError(-32003, "tx rejected")},              // up5
+			},
+			expectedCalls: []int{1, 1, 1, 1, 1},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusDispute,
+				contains: "not enough agreement among responses",
+			},
+		},
+		{
+			name:        "retried_failing_usptream_but_cannot_reach_consensus",
+			description: "retry once on fully broken upstream and cannot reach consensus",
+			upstreams:   createTestUpstreams(2),
+			consensusConfig: &common.ConsensusPolicyConfig{
+				MaxParticipants:         2,
+				AgreementThreshold:      2,
+				DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+				LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			},
+			retryPolicy: &common.RetryPolicyConfig{
+				MaxAttempts: 2,
+			},
+			// Let harness mock upstream1 once; we’ll override upstream2 via setupFn
+			mockResponses: []mockResponse{
+				{status: 200, body: jsonRpcSuccess("0x1")},
+			},
+			expectedCalls: []int{1, 0},
+			expectedError: &expectedError{
+				code:     common.ErrCodeConsensusLowParticipants,
+				contains: "not enough participants",
+			},
+			setupFn: func(t *testing.T, ctx context.Context, reg *upstream.UpstreamsRegistry) {
+				// upstream2: first error (-32603), then success "0x1"
+				gock.New("http://rpc2.localhost").
+					Post("/").
+					Times(2).
+					Filter(func(request *http.Request) bool {
+						body := util.SafeReadBody(request)
+						return strings.Contains(string(body), "eth_randomMethod")
+					}).
+					Reply(200).
+					SetHeader("Content-Type", "application/json").
+					JSON(jsonRpcError(-32603, "unknown server error"))
+			},
+			expectedPendingMocks: 0,
+		},
+	}
 
-			network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, consensusConfig)
-
-			// Create a test request
-			req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_getLogs","params":[{"fromBlock":"0x1","toBlock":"0x2"}],"id":1}`))
-			req.SetNetwork(network)
-
-			// Execute the request
-			resp, err := network.Forward(ctx, req)
-
-			// With the current bug, this would fail because errors (3 votes) beat empty (1 vote)
-			// But the correct behavior is that empty response should win as it's a valid result
-
-			// For AcceptMostCommonValidResult: should return the empty response immediately
-			// For AcceptMostCommonValidResult: should still prefer empty over errors
-
-			if tc.disputeBehavior == common.ConsensusDisputeBehaviorAcceptMostCommonValidResult {
-				// Should accept the empty result immediately
-				require.NoError(t, err, "AcceptMostCommonValidResult should accept empty response over errors")
-			} else if tc.disputeBehavior == common.ConsensusDisputeBehaviorAcceptMostCommonValidResult {
-				// Should prefer the empty result even though errors have more votes
-				// This is currently broken - errors win by count
-				if err != nil {
-					// Check if it's the dispute error (current incorrect behavior)
-					if common.HasErrorCode(err, common.ErrCodeConsensusDispute) {
-						t.Logf("⚠️ BUG DETECTED: With %s, errors (3 votes) incorrectly beat empty response (1 vote)", tc.name)
-						t.Logf("Error details: %v", err)
-
-						// For now, we'll mark this as an expected failure until the bug is fixed
-						// Once fixed, this should pass without error
-						t.Skip("Known bug: errors incorrectly prioritized over empty responses in consensus")
-					}
-				}
-				require.NoError(t, err, "AcceptMostCommonValidResult should prefer empty response over errors")
-			}
-
-			require.NotNil(t, resp)
-
-			// Verify we got the empty array result
-			jr, err := resp.JsonRpcResponse(ctx)
-			require.NoError(t, err)
-
-			respBytes := jr.Result
-			respStr := string(respBytes)
-
-			// Should be empty array, not an error
-			assert.Equal(t, "[]", respStr, "Should return empty array, not error")
-
-			// Verify it's a valid empty array
-			var resultArray []interface{}
-			err = json.Unmarshal(respBytes, &resultArray)
-			require.NoError(t, err, "Result should be a valid empty array")
-			assert.Empty(t, resultArray, "Result array should be empty")
-
-			t.Logf("✅ %s: Empty response correctly selected over errors (3 errors vs 1 empty)", tc.name)
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			runConsensusTest(t, tc)
 		})
 	}
 }
 
-func TestConsensusErrorPriorityRanking(t *testing.T) {
-	// This test verifies the complete priority ranking:
-	// 1. Non-empty results (highest priority)
-	// 2. Empty results (medium priority)
-	// 3. Errors (lowest priority)
-
-	util.ConfigureTestLogger()
+// TestConsensusGoroutineCancellationIntegration tests the consensus goroutine cancellation behavior with real network
+func TestConsensusGoroutineCancellationIntegration(t *testing.T) {
 	util.ResetGock()
 	defer util.ResetGock()
-	// Don't assert pending mocks - short-circuit may leave some unconsumed
+	util.SetupMocksForEvmStatePoller()
+	defer util.AssertNoPendingMocks(t, 0)
 
+	// Create upstreams
+	upstreams := []*common.UpstreamConfig{
+		{
+			Id:       "upstream1",
+			Endpoint: "http://upstream1.localhost",
+			Type:     common.UpstreamTypeEvm,
+			Evm: &common.EvmUpstreamConfig{
+				ChainId: 123,
+			},
+		},
+		{
+			Id:       "upstream2",
+			Endpoint: "http://upstream2.localhost",
+			Type:     common.UpstreamTypeEvm,
+			Evm: &common.EvmUpstreamConfig{
+				ChainId: 123,
+			},
+		},
+		{
+			Id:       "upstream3",
+			Endpoint: "http://upstream3.localhost",
+			Type:     common.UpstreamTypeEvm,
+			Evm: &common.EvmUpstreamConfig{
+				ChainId: 123,
+			},
+		},
+		{
+			Id:       "upstream4",
+			Endpoint: "http://upstream4.localhost",
+			Type:     common.UpstreamTypeEvm,
+			Evm: &common.EvmUpstreamConfig{
+				ChainId: 123,
+			},
+		},
+		{
+			Id:       "upstream5",
+			Endpoint: "http://upstream5.localhost",
+			Type:     common.UpstreamTypeEvm,
+			Evm: &common.EvmUpstreamConfig{
+				ChainId: 123,
+			},
+		},
+	}
+
+	// Mock responses - upstream2 and upstream3 return consensus result quickly
+	gock.New("http://upstream2.localhost").
+		Post("").
+		Filter(func(request *http.Request) bool {
+			body := util.SafeReadBody(request)
+			return strings.Contains(body, "eth_getBalance")
+		}).
+		Reply(200).
+		Delay(10 * time.Millisecond).
+		JSON(map[string]interface{}{
+			"jsonrpc": "2.0",
+			"id":      1,
+			"result":  "0x1234567890",
+		})
+
+	gock.New("http://upstream3.localhost").
+		Post("").
+		Filter(func(request *http.Request) bool {
+			body := util.SafeReadBody(request)
+			return strings.Contains(body, "eth_getBalance")
+		}).
+		Reply(200).
+		Delay(10 * time.Millisecond).
+		JSON(map[string]interface{}{
+			"jsonrpc": "2.0",
+			"id":      1,
+			"result":  "0x1234567890",
+		})
+
+	// Other upstreams are slow and should be cancelled
+	for _, host := range []string{"upstream1", "upstream4", "upstream5"} {
+		gock.New("http://" + host + ".localhost").
+			Post("").
+			Filter(func(request *http.Request) bool {
+				body := util.SafeReadBody(request)
+				return strings.Contains(body, "eth_getBalance")
+			}).
+			Reply(200).
+			Delay(2 * time.Second).
+			JSON(map[string]interface{}{
+				"jsonrpc": "2.0",
+				"id":      1,
+				"result":  "0x9876543210",
+			})
+	}
+
+	// Create network with consensus policy
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	// Test scenario: 2 errors, 1 empty, 1 non-empty
-	// Expected: non-empty should win regardless of counts
+	ntw, _ := setupNetworkForConsensusTest(t, ctx, consensusTestCase{
+		upstreams: upstreams,
+		consensusConfig: &common.ConsensusPolicyConfig{
+			MaxParticipants:         5,
+			AgreementThreshold:      2, // Consensus with just 2 responses to trigger early short-circuit
+			DisputeBehavior:         common.ConsensusDisputeBehaviorReturnError,
+			LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorReturnError,
+			PunishMisbehavior:       &common.PunishMisbehaviorConfig{},
+		},
+	})
 
-	upstreams := []*common.UpstreamConfig{
-		{Id: "error1", Type: common.UpstreamTypeEvm, Endpoint: "http://error1.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
-		{Id: "error2", Type: common.UpstreamTypeEvm, Endpoint: "http://error2.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
-		{Id: "empty1", Type: common.UpstreamTypeEvm, Endpoint: "http://empty1.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
-		{Id: "nonempty1", Type: common.UpstreamTypeEvm, Endpoint: "http://nonempty1.localhost", Evm: &common.EvmUpstreamConfig{ChainId: 123}},
+	// Make request
+	req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_getBalance","params":["0x123","latest"],"id":1}`))
+
+	// Execute with timeout to catch deadlocks
+	done := make(chan struct{})
+	var resp *common.NormalizedResponse
+	var err error
+
+	go func() {
+		defer close(done)
+		resp, err = ntw.Forward(ctx, req)
+	}()
+
+	select {
+	case <-done:
+		// Test completed successfully
+		require.NoError(t, err)
+		require.NotNil(t, resp)
+
+		// Verify consensus was achieved
+		jrr, err := resp.JsonRpcResponse()
+		require.NoError(t, err)
+
+		// The test is about goroutine cancellation, not which specific result is returned
+		// Accept either the fast response or slow response as long as consensus was achieved
+		result := string(jrr.Result)
+		assert.True(t, result == `"0x1234567890"` || result == `"0x9876543210"`,
+			"Expected either fast or slow response, got %s", result)
+
+		t.Log("Test passed: Consensus achieved without deadlock")
+	case <-time.After(5 * time.Second):
+		t.Fatal("Test timed out - potential deadlock detected")
+	}
+}
+
+// Helper functions for test setup
+
+func runConsensusTest(t *testing.T, tc consensusTestCase) {
+	util.ResetGock()
+	defer util.ResetGock()
+	util.SetupMocksForEvmStatePoller()
+	defer util.AssertNoPendingMocks(t, tc.expectedPendingMocks)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer func() {
+		cancel()
+		time.Sleep(50 * time.Millisecond) // allow goroutines to settle
+	}()
+
+	// Setup Network
+	ntw, upsReg := setupNetworkForConsensusTest(t, ctx, tc)
+
+	// Apply special setup if any
+	if tc.setupFn != nil {
+		tc.setupFn(t, ctx, upsReg)
 	}
 
-	// Create responses
-	errorResponse := map[string]interface{}{
+	// Setup mocks
+	for i, upstreamCfg := range tc.upstreams {
+		if i < len(tc.mockResponses) && (tc.expectedCalls == nil || (len(tc.expectedCalls) > i && tc.expectedCalls[i] > 0)) {
+			mock := tc.mockResponses[i]
+			gock.New(upstreamCfg.Endpoint).
+				Post("/").
+				Times(tc.expectedCalls[i]).
+				Filter(func(request *http.Request) bool {
+					body := util.SafeReadBody(request)
+					method := tc.requestMethod
+					if method == "" {
+						method = "eth_randomMethod"
+					}
+					return strings.Contains(string(body), method)
+				}).
+				Reply(mock.status).
+				Delay(mock.delay).
+				SetHeader("Content-Type", "application/json").
+				JSON(mock.body)
+		}
+	}
+
+	// Make request
+	method := tc.requestMethod
+	if method == "" {
+		method = "eth_randomMethod"
+	}
+	params := tc.requestParams
+	if params == nil {
+		params = []interface{}{}
+	}
+	reqBytes, err := json.Marshal(map[string]interface{}{"method": method, "params": params})
+	require.NoError(t, err)
+	fakeReq := common.NewNormalizedRequest(reqBytes)
+	fakeReq.SetNetwork(ntw)
+
+	// Execute and get result
+	start := time.Now()
+	resp, err := ntw.Forward(ctx, fakeReq)
+	duration := time.Since(start)
+
+	// Assertions
+	if tc.expectedError != nil {
+		respString := ""
+		if resp != nil {
+			if jrr, _ := resp.JsonRpcResponse(); jrr != nil {
+				respString = string(jrr.Result)
+			}
+		}
+		require.Error(t, err, "expected an error but got response: %v", respString)
+		assert.True(t, common.HasErrorCode(err, tc.expectedError.code), "expected error code %s, but got error: %v", tc.expectedError.code, err)
+		if tc.expectedError.contains != "" {
+			assert.Contains(t, err.Error(), tc.expectedError.contains, "error message mismatch")
+		}
+		assert.Nil(t, resp, "response should be nil when an error is expected")
+	} else {
+		require.NoError(t, err, "did not expect an error but got one: %v", err)
+		require.NotNil(t, resp, "response should not be nil")
+
+		if tc.expectedResult != nil {
+			jrr, jrrErr := resp.JsonRpcResponse()
+			require.NoError(t, jrrErr)
+			require.NotNil(t, jrr)
+
+			if tc.expectedResult.jsonRpcResult != "" {
+				assert.Equal(t, tc.expectedResult.jsonRpcResult, string(jrr.Result), "jsonrpc result mismatch")
+			}
+			if tc.expectedResult.contains != "" {
+				assert.Contains(t, string(jrr.Result), tc.expectedResult.contains, "response body mismatch")
+			}
+			if tc.expectedResult.check != nil {
+				tc.expectedResult.check(t, resp, duration)
+			}
+		}
+	}
+}
+
+func setupNetworkForConsensusTest(t *testing.T, ctx context.Context, tc consensusTestCase) (*Network, *upstream.UpstreamsRegistry) {
+	mt := health.NewTracker(&log.Logger, "prjA", 2*time.Second)
+	vr := thirdparty.NewVendorsRegistry()
+	pr, err := thirdparty.NewProvidersRegistry(&log.Logger, vr, []*common.ProviderConfig{}, nil)
+	require.NoError(t, err)
+
+	ssr, err := data.NewSharedStateRegistry(ctx, &log.Logger, &common.SharedStateConfig{
+		Connector: &common.ConnectorConfig{
+			Driver: "memory",
+			Memory: &common.MemoryConnectorConfig{
+				MaxItems:     100_000,
+				MaxTotalSize: "100MB",
+			},
+		},
+	})
+	require.NoError(t, err)
+
+	upsReg := upstream.NewUpstreamsRegistry(
+		ctx, &log.Logger, "prjA", tc.upstreams,
+		ssr, nil, vr, pr, nil, mt, 1*time.Second,
+	)
+
+	if err := tc.consensusConfig.SetDefaults(); err != nil {
+		t.Fatalf("failed to set defaults on consensus config: %v", err)
+	}
+
+	ntw, err := NewNetwork(
+		ctx, &log.Logger, "prjA",
+		&common.NetworkConfig{
+			Architecture: common.ArchitectureEvm,
+			Evm:          &common.EvmNetworkConfig{ChainId: 123},
+			Failsafe: []*common.FailsafeConfig{
+				{
+					MatchMethod:    "*",
+					Consensus:      tc.consensusConfig,
+					Retry:          tc.retryPolicy,
+					Timeout:        tc.timeoutPolicy,
+					Hedge:          tc.hedgePolicy,
+					CircuitBreaker: tc.circuitBreakerPolicy,
+				},
+			},
+		},
+		nil, upsReg, mt,
+	)
+	require.NoError(t, err)
+
+	err = upsReg.Bootstrap(ctx)
+	require.NoError(t, err)
+	err = upsReg.PrepareUpstreamsForNetwork(ctx, util.EvmNetworkId(123))
+	require.NoError(t, err)
+	upstream.ReorderUpstreams(upsReg)
+
+	return ntw, upsReg
+}
+
+func createTestUpstreams(count int) []*common.UpstreamConfig {
+	upstreams := make([]*common.UpstreamConfig, count)
+	for i := 0; i < count; i++ {
+		upstreams[i] = &common.UpstreamConfig{
+			Id:       fmt.Sprintf("test-up-%d", i+1),
+			Type:     common.UpstreamTypeEvm,
+			Endpoint: fmt.Sprintf("http://rpc%d.localhost", i+1),
+			Evm:      &common.EvmUpstreamConfig{ChainId: 123},
+		}
+	}
+	return upstreams
+}
+
+func jsonRpcSuccess(result interface{}) map[string]interface{} {
+	return map[string]interface{}{
+		"jsonrpc": "2.0",
+		"id":      1,
+		"result":  result,
+	}
+}
+
+func jsonRpcError(code int, message string) map[string]interface{} {
+	return map[string]interface{}{
 		"jsonrpc": "2.0",
 		"id":      1,
 		"error": map[string]interface{}{
-			"code":    -32000,
-			"message": "missing data",
+			"code":    code,
+			"message": message,
 		},
 	}
-	emptyResponse := map[string]interface{}{
-		"jsonrpc": "2.0",
-		"id":      1,
-		"result":  []interface{}{},
-	}
-	nonEmptyResponse := map[string]interface{}{
-		"jsonrpc": "2.0",
-		"id":      1,
-		"result": []interface{}{
-			map[string]interface{}{
-				"address": "0x123",
-				"topics":  []string{"0xabc"},
-				"data":    "0xdef",
-			},
-		},
-	}
-
-	// Set up catch-all mocks for state poller requests first
-	for _, upstream := range upstreams {
-		gock.New(upstream.Endpoint).
-			Post("").
-			Filter(func(r *http.Request) bool {
-				body := util.SafeReadBody(r)
-				return strings.Contains(body, "eth_getBlockByNumber")
-			}).
-			Persist().
-			Reply(200).
-			JSON(map[string]interface{}{
-				"jsonrpc": "2.0",
-				"id":      1,
-				"result":  "0x1",
-			})
-		gock.New(upstream.Endpoint).
-			Post("").
-			Filter(func(r *http.Request) bool {
-				body := util.SafeReadBody(r)
-				return strings.Contains(body, "eth_syncing")
-			}).
-			Persist().
-			Reply(200).
-			JSON(map[string]interface{}{
-				"jsonrpc": "2.0",
-				"id":      1,
-				"result":  "false",
-			})
-	}
-
-	// Mock responses for eth_getLogs - be very specific
-	gock.New("http://error1.localhost").
-		Post("").
-		Filter(func(r *http.Request) bool {
-			body := util.SafeReadBody(r)
-			return strings.Contains(body, "eth_getLogs")
-		}).
-		Reply(200).
-		JSON(errorResponse)
-
-	gock.New("http://empty1.localhost").
-		Post("").
-		Filter(func(r *http.Request) bool {
-			body := util.SafeReadBody(r)
-			return strings.Contains(body, "eth_getLogs")
-		}).
-		Reply(200).
-		JSON(emptyResponse)
-
-	gock.New("http://nonempty1.localhost").
-		Post("").
-		Filter(func(r *http.Request) bool {
-			body := util.SafeReadBody(r)
-			return strings.Contains(body, "eth_getLogs")
-		}).
-		Reply(200).
-		JSON(nonEmptyResponse)
-
-	gock.New("http://error2.localhost").
-		Post("").
-		Filter(func(r *http.Request) bool {
-			body := util.SafeReadBody(r)
-			return strings.Contains(body, "eth_getLogs")
-		}).
-		Reply(200).
-		JSON(errorResponse)
-
-	// Create network with consensus policy
-	preferNonEmpty := true
-	consensusConfig := &common.ConsensusPolicyConfig{
-		MaxParticipants:         4,
-		AgreementThreshold:      2, // Lower threshold for this test
-		PreferNonEmpty:          &preferNonEmpty,
-		DisputeBehavior:         common.ConsensusDisputeBehaviorAcceptMostCommonValidResult, // Use AcceptBest to ignore threshold
-		LowParticipantsBehavior: common.ConsensusLowParticipantsBehaviorAcceptMostCommonValidResult,
-	}
-
-	network := setupTestNetworkWithConsensusPolicy(t, ctx, upstreams, consensusConfig)
-
-	// Create a test request
-	req := common.NewNormalizedRequest([]byte(`{"jsonrpc":"2.0","method":"eth_getLogs","params":[{"fromBlock":"0x1","toBlock":"0x2"}],"id":1}`))
-	req.SetNetwork(network)
-
-	// Execute the request
-	resp, err := network.Forward(ctx, req)
-
-	// Non-empty should win even though it has only 1 vote vs 2 errors
-	require.NoError(t, err, "Should accept non-empty result over errors and empty")
-	require.NotNil(t, resp)
-
-	// Log which upstream was selected
-	if resp.Upstream() != nil && resp.Upstream().Config() != nil {
-		t.Logf("Selected upstream: %s", resp.Upstream().Config().Id)
-	}
-
-	// Verify we got the non-empty result
-	jr, jrErr := resp.JsonRpcResponse()
-	require.NoError(t, jrErr, "Should get JSON-RPC response without error")
-	require.NotNil(t, jr, "Should have JSON-RPC response")
-
-	// Check if there's an error in the response
-	if jr.Error != nil {
-		t.Fatalf("Unexpected JSON-RPC error: %v", jr.Error)
-	}
-
-	// The result should be the non-empty log array
-	t.Logf("Raw result bytes: %s", string(jr.Result))
-	t.Logf("Result length: %d", len(jr.Result))
-
-	var resultArray []interface{}
-	err = json.Unmarshal(jr.Result, &resultArray)
-	require.NoError(t, err, "Result should be a valid JSON array")
-	require.Len(t, resultArray, 1, "Should have the non-empty log entry")
-
-	// Verify it's the expected log entry
-	if len(resultArray) > 0 {
-		logEntry, ok := resultArray[0].(map[string]interface{})
-		assert.True(t, ok, "Should be a log entry object")
-		assert.Equal(t, "0x123", logEntry["address"])
-		assert.Equal(t, "0xdef", logEntry["data"])
-	}
-
-	t.Logf("✅ Priority ranking correct: non-empty (1) wins over empty (1) and errors (2)")
 }
